% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{add_class}
\alias{add_class}
\title{Add a class to an object}
\usage{
add_class(x, class)
}
\arguments{
\item{x}{An object}

\item{class}{The class to add to the object.}
}
\value{
The object \code{x} that will inherit the class \code{class}.
}
\description{
Add a class to an object
}
\examples{
\donttest{
library(metan)
df <-
data_ge2 \%>\%
add_class("my_class")
class(df)
}

}
