% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange_ggplot.R
\name{arrange_ggplot}
\alias{arrange_ggplot}
\title{Arrange multiple ggplot2 graphics in a single image window}
\usage{
arrange_ggplot(..., nrow = NULL, ncol = NULL)
}
\arguments{
\item{...}{An object of class \code{gg}}

\item{nrow}{The number of rows}

\item{ncol}{The number of columns}
}
\value{
None.
}
\description{
This is a helper function to arrange ggplot2 objects in the metan package
}
\examples{
library(ggplot2)
p1 <- ggplot(mtcars, aes(wt, mpg)) +
      geom_point()

p2 <- ggplot(mpg, aes(class, hwy)) +
             geom_boxplot()

arrange_ggplot(p1, p2)

}
