% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_cv.R
\name{bind_cv}
\alias{bind_cv}
\title{Bind cross-validation objects.}
\usage{
bind_cv(...)
}
\arguments{
\item{...}{Input objects of class \code{cv_ammi}, \code{cv_ammif} or
\code{cv_blup}.}
}
\description{
Helper function that combines objects of class \code{cv_ammi}, \code{cv_ammif}
 or \code{cv_blup}. It is useful when looking for a boxplot containing
the RMSPD values of those cross-validation procedures.
}
\examples{

\dontrun{
library(METAAB)
AMMI = cv_ammif(data_ge,
                        resp = GY,
                        gen = GEN,
                        env = ENV,
                        rep = REP,
                        nboot = 100,
                        nrepval = 2)
BLUP = cv_blup(data_ge,
                        resp = GY,
                        gen = GEN,
                        env = ENV,
                        rep = REP,
                        nboot = 100,
                        nrepval = 2)
bind_data = bind_cv(AMMI, BLUP)
plot(bind_data)
}

}
\author{
Tiago Olivoto \email{tiagoolivoto@gmail.com}
}
