% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blup_indexes.R
\name{blup_indexes}
\alias{blup_indexes}
\alias{hmgv}
\alias{rpgv}
\alias{hmrpgv}
\title{Stability indexes based on a mixed-effect model}
\usage{
hmgv(model)

rpgv(model)

hmrpgv(model)

blup_indexes(model)
}
\arguments{
\item{model}{An object of class \code{waasb} computed with \code{\link[=waasb]{waasb()}} or
\code{\link[=gamem_met]{gamem_met()}}.}
}
\value{
A tibble containing the indexes.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
\itemize{
\item \code{\link[=hmgv]{hmgv()}} Computes the harmonic mean of genotypic values (HMGV).
\item \code{\link[=rpgv]{rpgv()}} Computes the relative performance of the genotypic values (RPGV).
\item \code{\link[=hmrpgv]{hmrpgv()}} Computes the harmonic mean of the relative performance of
genotypic values (HMRPGV).
\item \code{\link[=blup_indexes]{blup_indexes()}} Is a wrapper around the above functions that also computes
the WAASB index (Olivoto et al. 2019) if an object computed with \code{\link[=waasb]{waasb()}} is
used as input data.
}
}
\details{
The indexes computed with this function have been used to select genotypes
with stability performance in a mixed-effect model framework. Some examples
are in Alves et al (2018), Azevedo Peixoto et al. (2018), Dias et al. (2018)
and Colombari Filho et al. (2013).

\loadmathjax
The HMGV index is computed as
\mjsdeqn{HMG{V_i} = \frac{E}{{\sum\limits_{j = 1}^E {\frac{1}{{G{v_{ij}}}}}}}}

where \mjseqn{E} is the number of environments included in the analysis,
\mjseqn{Gv_{ij}} is the genotypic value (BLUP) for the ith genotype in the
jth environment.

The RPGV index is computed as
\mjsdeqn{RPGV_i = \frac{1}{E}{\sum\limits_{j = 1}^E {Gv_{ij}} /\mathop \mu
\nolimits_j }}

The HMRPGV index is computed as
\mjsdeqn{HMRPG{V_i} = \frac{E}{{\sum\limits_{j = 1}^E
{\frac{1}{{G{v_{ij}}/{\mu _j}}}} }}}
}
\examples{
\donttest{
library(metan)
res_ind <- waasb(data_ge,
                 env = ENV,
                 gen = GEN,
                 rep = REP,
                 resp = c(GY, HM),
                 verbose = FALSE)
model_indexes <- blup_indexes(res_ind)
gmd(model_indexes)
}


}
\references{
Alves, R.S., L. de Azevedo Peixoto, P.E. Teodoro, L.A. Silva, E.V.
Rodrigues, M.D.V. de Resende, B.G. Laviola, and L.L. Bhering. 2018.
Selection of Jatropha curcas families based on temporal stability and
adaptability of genetic values. Ind. Crops Prod. 119:290-293.
\doi{10.1016/J.INDCROP.2018.04.029}

Azevedo Peixoto, L. de, P.E. Teodoro, L.A. Silva, E.V. Rodrigues, B.G.
Laviola, and L.L. Bhering. 2018. Jatropha half-sib family selection with
high adaptability and genotypic stability. PLoS One 13:e0199880.
\doi{10.1371/journal.pone.0199880}

Colombari Filho, J.M., M.D.V. de Resende, O.P. de Morais, A.P. de Castro,
E.P. Guimaraes, J.A. Pereira, M.M. Utumi, and F. Breseghello. 2013. Upland
rice breeding in Brazil: a simultaneous genotypic evaluation of stability,
adaptability and grain yield. Euphytica 192:117-129.
\doi{10.1007/s10681-013-0922-2}

Dias, P.C., A. Xavier, M.D.V. de Resende, M.H.P. Barbosa, F.A. Biernaski,
R.A. Estopa. 2018. Genetic evaluation of Pinus taeda clones from somatic
embryogenesis and their genotype x environment interaction. Crop Breed.
Appl. Biotechnol. 18:55-64.
\doi{10.1590/1984-70332018v18n1a8}

Olivoto, T., A.D.C. L{\'{u}}cio, J.A.G. da silva, V.S. Marchioro, V.Q. de
Souza, and E. Jost. 2019. Mean performance and stability in
multi-environment trials I: Combining features of AMMI and BLUP techniques.
Agron. J. 111:2949-2960. \doi{10.2134/agronj2019.03.0220}

Resende MDV (2007) Matematica e estatistica na analise de experimentos e no
melhoramento genetico. Embrapa Florestas, Colombo
}
\author{
Tiago Olivoto \email{tiagoolivoto@gmail.com}
}
