% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_coef.R
\name{corr_focus}
\alias{corr_focus}
\title{Focus on section of a correlation matrix}
\usage{
corr_focus(model, ...)
}
\arguments{
\item{model}{A model computed with \code{\link[=corr_coef]{corr_coef()}} or a symmetric matrix, often
produced with \code{\link[stats:cor]{stats::cor()}}.}

\item{...}{One or more unquoted variable name separated by commas. Variable
names can be used as if they were positions in the data frame, so
expressions like \code{x:y} can be used to select a range of variables.}
}
\value{
A tibble
}
\description{
Select a set of variables from a correlation matrix to keep as the columns,
and exclude these or all other variables from the rows.
}
\examples{
corr_coef(data_ge2) |> corr_focus(PH)
}
