% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_blup.R
\name{cv_blup}
\alias{cv_blup}
\title{Cross-validation procedure}
\usage{
cv_blup(.data, env, gen, rep, resp, nboot = 100, nrepval,
  random = "gen", verbose = TRUE)
}
\arguments{
\item{.data}{The dataset containing the columns related to Environments,
Genotypes, replication/block and response variable(s).}

\item{env}{The name of the column that contains the levels of the
environments.}

\item{gen}{The name of the column that contains the levels of the genotypes.}

\item{rep}{The name of the column that contains the levels of the
replications/blocks.}

\item{resp}{The response variable.}

\item{nboot}{The number of resamples to be used in the cross-validation. Defaults to 100.}

\item{nrepval}{The number of replicates (r) from total number of replicates
(R) to be used in the modeling dataset. Only one replicate is used as
validating data each step, so, \code{Nrepval} must be equal \code{R-1}}

\item{random}{The effects of the model assumed to be random. Default is
\code{random = 'gen'} (genotype and genotype-vs-environment as random
effects. Other values allowed are \code{random = 'env'} (environment,
genotype-vs-environment and block-within-environment random effects) or
\code{random = 'all'} all effects except the intercept are assumed to be
random effects.}

\item{verbose}{A logical argument to define if a progress bar is shown.
Default is \code{TRUE}.}
}
\value{
An object of class \code{cv_blup} with the following items:
\item{RMSPD}{A vector with nboot-estimates of the root mean squared
prediction difference between predicted and validating data.}
\item{RMSPDmean}{The mean of RMSPDmean estimates.}
}
\description{
Cross-validation for blup prediction.
}
\details{
This function provides a cross-validation procedure for mixed models using
replicate-based data. By default, complete blocks are randomly selected
within each environment. In each iteration, the original dataset is split up
into two datasets: training and validation data. The 'training' set has all
combinations (genotype x environment) with the number of replications
informed in \code{nrepval}. The 'validation' set has the remaining
replication. The estimated values are compared with the 'validation' data
and the Root Means Square Prediction Difference is computed. At the end of
boots, a list is returned.
}
\examples{

\dontrun{
library(metan)
model = cv_blup(data_ge,
                        env = ENV,
                        gen = GEN,
                        rep = REP,
                        resp = GY,
                        nboot = 100,
                        nrepval = 2)

# Alternatively using the pipe operator \%>\%
model = data_ge \%>\%
        cv_blup(ENV, GEN, REP, GY, 100, 2)

}

}
\seealso{
\code{\link{cv_ammi}}, \code{\link{cv_ammif}}
}
\author{
Tiago Olivoto \email{tiagoolivoto@gmail.com}
}
