\name{data_ge}
\alias{data_ge}
\title{A replicate-based data of 10 genotypes cultivated in 14 environments}

\description{

 This dataset contain data on two variables assessed in 10 genotypes growing in in 11 environments. The experimental design was a RCBD with 3 replicates(blocks). This data provide examples for several functions of \pkg{METAAB} package.
}

\arguments{
The following collumns in \code{data_ge} are:

  \item{ENV}{A factor with 14 levels; each level represents one cultivation environment.}

  \item{GEN}{A factor with 10 levels; each level represents one genotype.}

  \item{REP}{A factor with 3 levels; each level represents one replication/block.}

  \item{GY}{A continuous variable (grain yield) observed in each plot.}

  \item{HM}{A continuous variable (hectoliter mass) observed in each plot.}

  }
\author{Tiago Olivoto \email{tiagoolivoto@gmail.com}}

