\name{data_ge2}
\alias{data_ge2}
\title{A replicate-based data of 13 maize hybrids cultivated in 4 environments}

\description{

 This dataset contain data on 15 traits assessed in 13 maize hybrids growing in 04 environments. The experimental design was a RCBD with 3 blocks and 5 replications per block. This data provide examples for several functions of \pkg{METAAB} package.
}

\arguments{
The following collumns are in \code{data_ge}:

  \item{ENV}{A factor with 04 levels; each level represents one cultivation environment.}
  \item{GEN}{A factor with 13 levels; each level represents one maize hybrid.}
  \item{REP}{A factor with 3 levels; each level represents one replication/block.}
  \item{PLA}{A factor with 5 levels; each level represents one plant assessed within the block.}
  \item{PH}{Plant height, in cm.}
  \item{EH}{Ear height, in cm.}
  \item{EP}{Ear position, i.e., the ratio EH/PH.}
  \item{EL}{Ear length, in cm.}
  \item{ED}{Ear diameter, in mm.}
  \item{CL}{Cob length, in cm.}
  \item{CD}{Cob diameter, in mm.}
  \item{CW}{Cob weight, in g.}
  \item{KW}{Kernel weight, in cm.}
  \item{NR}{Number of rows.}
  \item{NKR}{Number of kernels per row.}
  \item{CDED}{Cob diameter / Ear diameter ratio.}
  \item{PERK}{Percentage of kernels.}
  \item{TKW}{Thousand-kernel weight}
  \item{NKE}{Number of kernels per row.}
  }
\author{Tiago Olivoto \email{tiagoolivoto@gmail.com}}

