% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_stat.R
\name{desc_stat}
\alias{desc_stat}
\title{Descriptive statistics}
\usage{
desc_stat(.data = NULL, ..., values = NULL, level = 0.95,
  digits = 4)
}
\arguments{
\item{.data}{The data to be analyzed. Must be a dataframe or an object of
class \code{split_factors}.}

\item{...}{A single variable name or a comma-separated list of unquoted variables names.}

\item{values}{An alternative way to pass the data to the function. It must be a numeric
vector.}

\item{level}{The confidence level to compute the confidence interval of mean. Defaults to 0.95.}

\item{digits}{The number of significant figures.}
}
\description{
Compute the most used measures of central tendency, position, and dispersion.
}
\examples{
\dontrun{

library(metan)
library(dplyr)

data_ge2 \%>\%
  desc_stat(TKW)

vect = data_ge2 \%>\%
  select(TKW) \%>\%
  pull()

desc_stat(values = vect)

data_ge2 \%>\%
  split_factors(ENV) \%>\%
  desc_stat(EP, EL, PH, CL, CW, NR, NKR)
}

}
\author{
Tiago Olivoto \email{tiagoolivoto@gmail.com}
}
