% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ge_plot.R
\name{ge_plot}
\alias{ge_plot}
\title{Graphical analysis of genotype-vs-environment interaction}
\usage{
ge_plot(.data, env, gen, resp, theme = theme_waasb())
}
\arguments{
\item{.data}{The dataset containing the columns related to Environments,
Genotypes, replication/block and response variable(s).}

\item{env}{The name of the column that contains the levels of the
environments}

\item{gen}{The name of the column that contains the levels of the genotypes.}

\item{resp}{The response variable.}

\item{theme}{The graphical theme of the plot. Default is \code{theme =
theme_waasb()}. Please, see \code{?WAASB::theme_waasb}. An own theme can be
applied using the arguments: \code{theme = theme(some stuff
here)}. For more details, please, see \code{?ggplot2::theme}.}
}
\description{
This function produces a line plot for a graphical interpretation of the
genotype-vs-environment interaction. By default, environments are in the
x axis whereas the genotypes are depicted by different lines. The y axis
contains the value of the selected variable.
}
\examples{
library(metan)
library(ggplot2)
ge_plot(data_ge2, ENV, GEN, PH)
ge_plot(data_ge2, ENV, GEN, PH)+
theme_gray()
}
\author{
Tiago Olivoto \email{tiagoolivoto@gmail.com}
}
