% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ge_stats.R
\name{ge_stats}
\alias{ge_stats}
\title{Statistics for genotype-vs-environment interaction}
\usage{
ge_stats(.data, env, gen, rep, resp, verbose = TRUE, prob = 0.05)
}
\arguments{
\item{.data}{The dataset containing the columns related to Environments,
Genotypes, replication/block and response variable(s).}

\item{env}{The name of the column that contains the levels of the
environments.}

\item{gen}{The name of the column that contains the levels of the genotypes.}

\item{rep}{The name of the column that contains the levels of the
replications/blocks.}

\item{resp}{The response variable(s). To analyze multiple variables in a
single procedure use, for example, \code{resp = c(var1, var2, var3)}.}

\item{verbose}{Logical argument. If \code{verbose = FALSE} the code will run
silently.}

\item{prob}{The probability error assumed.}
}
\value{
An object of class \code{ge_stats} with the following items for each
variable:
\itemize{
\item \strong{individual} The individual analysis of variance. Call the function
\code{\link{anova_ind}} internally.
\item \strong{gai} The Geometric adaptability index. Call the function
\code{\link{gai}} internally.
\item \strong{ge_mean}, ge_effect The genotype-vs-environment means and effects, respectively.
\item \strong{ge_effect} The genotype-environment effects. Call the function
\code{\link{ge_effects}} internally.
\item \strong{gge_effect} The genotype plus genotype-vs-environment effects. Call
the function \code{\link{ge_effects}} internally.
\item \strong{ge_stats} The variance, sum of squares and mean squares for each
genotype.
\item \strong{ER} The Eberhart and Russell regression model. Call the function
\code{\link{ge_reg}} internally.
\item \strong{ANN} The Annicchiarico's genotypic confidence index. Call the
function \code{\link{Annicchiarico}} internally.
\item \strong{Ecoval} The Wrike's ecovalence. Call the function
\code{\link{ecovalence}} internally.
\item \strong{Shukla} The Shukla's stability variance. Call the function
\code{\link{Shukla}} internally.
\item \strong{Fox} The Fox's stability function. Call the function
\code{\link{Fox}} internally.
\item \strong{Superiority} The Lin and Binns' superiority index. Call the
function \code{\link{superiority}} internally.
}
}
\description{
Computes \strong{(i)} within-environment analysis of variance, GEI effect,
GEI means, and genotype plus GEI effects; \strong{(ii)} parametric statistics
including Annicchiarico's genotypic confidence index (1992), Ecovalence
(Wricke, 1965), regression-based stability (Eberhart and Russell., 1966),
Shukla's stability variance parameter (1972); and \strong{(iii)}
nonparametric Fox's stability function (Fox et al. 1990), and  superiority
index (Lin and Binns, 1988)
}
\examples{

library(metan)

model <- ge_stats(.data = data_ge2,
                  env = ENV,
                  gen = GEN,
                  rep = REP,
                  resp = PH)

#Alternatively, using the forward-pipe operator \%>\%
#More than one trait

model2 <- data_ge2 \%>\%
          ge_stats(ENV, GEN, REP, c(PH, EH, EL))


}
\references{
Annicchiarico, P. 1992. Cultivar adaptation and recommendation from alfalfa
trials in Northern Italy. Journal of Genetic & Breeding, 46:269-278

Eberhart, S.A., and W.A. Russell. 1966. Stability parameters for
comparing Varieties. Crop Sci. 6:36-40.
\href{https://www.crops.org/publications/cs/abstracts/6/1/CS0060010036}{doi:10.2135/cropsci1966.0011183X000600010011x}

Fox, P.N., B. Skovmand, B.K. Thompson, H.J. Braun, and R.
Cormier. 1990. Yield and adaptation of hexaploid spring triticale.
Euphytica 47:57-64.
\href{https://link.springer.com/article/10.1007/BF00040364}{doi:10.1007/BF00040364}.

Kang, M.S., and H.N. Pham. 1991. Simultaneous Selection for High
Yielding and Stable Crop Genotypes. Agron. J. 83:161.
\href{https://dl.sciencesocieties.org/publications/aj/abstracts/83/1/AJ0830010161}{doi:10.2134/agronj1991.00021962008300010037x}.

Lin, C.S., and M.R. Binns. 1988. A superiority measure of cultivar
performance for cultivar x location data. Can. J. Plant Sci. 68:193-198.
\href{http://pubs.aic.ca/doi/abs/10.4141/cjps88-018}{doi:10.4141/cjps88-018}

Shahbazi, E. 2019. Genotype selection and stability analysis for
seed yield of Nigella sativa using parametric and non-parametric
statistics. Sci. Hortic. (Amsterdam). 253:172-179.
\href{https://www.sciencedirect.com/science/article/pii/S0304423819303012}{doi:10.1016/j.scienta.2019.04.047}.

Shukla, G.K. 1972. Some statistical aspects of partitioning
genotype-environmental components of variability. Heredity. 29:238-245.
\href{http://www.nature.com/articles/hdy197287}{doi:10.1038/hdy.1972.87}.

Wricke, G. 1965. Zur berechnung der okovalenz bei sommerweizen und hafer.
Z. Pflanzenzuchtg 52:127-138.
}
\author{
Tiago Olivoto \email{tiagoolivoto@gmail.com}
}
