% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_data.R
\name{get_model_data}
\alias{get_model_data}
\title{Easily get data from a model}
\usage{
get_model_data(x, what = "Y", type = "GEN")
}
\arguments{
\item{x}{An object created with the functions \code{\link{waas}} or \code{\link{waasb}}.}

\item{what}{What should be captured from the model. The following options are allowed:
\code{"blupg"} for genotype's predicted mean and \code{"blupge"} for genotype-vs-environment's
predicted mean (if the object is of class \code{waasb} only); \code{"Y"} for raw means (Default);
\code{"PCn"} for the scores of the nth Interaction Principal Component Axis (IPCA),
 where \code{n} is the desired IPCA (e.g. \code{"PC3"}); \code{"WAASB"} for the weighted
 average of the absolute scores; \code{"PctResp"} and \code{"PctWAASB"} for the rescaled
 values of the response variable and WAASB, respectively; \code{"wResp"} and \code{"wWAASB"}
 for weight of the response variable and stability, respectively; \code{"OrResp"} and
 \code{"OrWAASB"} for the ranking regarding the response variable and WAASB, respectively;
 \code{"OrPC1"} for the ranking regarding the first principal component axix; \code{"WAASBY"}
 for the superiority index, \code{"OrWAASBY"} for the ranking regarding the superiority
 index.}

\item{type}{Chose if the statistics must be show by genotype (\code{type = "GEN"}, default)
or environment (\code{type = "ENV"}), when possible.}
}
\description{
Easily get data from a model of class \code{waas} or \code{waasb} (Olivoto et al., 2019a, 2019b)
}
\examples{

library(metan)

model = waas(data_ge2,
             env = ENV,
             gen = GEN,
             rep = REP,
             resp = c(PH, ED, TKW, NKR))
get_model_data(model, what = "OrWAASBY")


waasb(data_ge2, ENV, GEN, REP,
      resp = c(PH, ED, TKW, NKR)) \%>\%
get_model_data(what = "blupg")


}
\references{
Olivoto, T., A.D.C. L{\'{u}}cio, J.A.G. da silva, V.S. Marchioro,
 V.Q. de Souza, and E. Jost. 2019a. Mean performance and stability in multi-environment
  trials I: Combining features of AMMI and BLUP techniques. Agron. J. (in press).

Olivoto, T., A.D.C. L{\'{u}}cio, J.A.G. da silva, B.G. Sari, and M.I. Diel. 2019b.
 Mean performance and stability in multi-environment trials II: Selection based on multiple
  traits. Agron. J. (in press)
}
\author{
Tiago Olivoto \email{tiagoolivoto@gmail.com}
}
