\name{int.effects}
\alias{int.effects}
\title{Dataset with genotype-by-environment interaction effects}

\description{
This dataset contains a matrix with interaction effects of 10 genotypes cultivated in 5 environments.

}

\seealso{\code{\link{meansGxE}}}

\author{Tiago Olivoto \email{tiagoolivoto@gmail.com}}
