\name{meansGxE}
\alias{meansGxE}
\title{A dataset with means of 10 genotypes cultivated in 5 environments }

\description{
This dataset contains the means for grain yield of 10 genotypes cultivated in 5 environments. The interaction effects for this data is found in \link{int.effects}

}


\author{Tiago Olivoto \email{tiagoolivoto@gmail.com}}
