% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtsi.R
\name{mtsi}
\alias{mtsi}
\title{Multi-trait stability index}
\usage{
mtsi(.data, index = "waasb", SI = 15, mineval = 1, verbose = TRUE)
}
\arguments{
\item{.data}{An object of class \code{waasb} or \code{waas}.}

\item{index}{If \code{index = "waasb"} (Default) the multitrait index will
be computed considering the stability of genotypes only. If \code{index =
"waasby"} both stability and mean performance are considered. More details
can be seen in \code{\link{waasb}} and \code{\link{waas}} functions.}

\item{SI}{An integer [0-100]. The selection intensity in percentage of the
total number of genotypes.}

\item{mineval}{The minimum value so that an eigenvector is retained in the
factor analysis.}

\item{verbose}{If \code{verbose = TRUE} (Default) then some results are
shown in the console.}
}
\description{
Multi-trait stability index in analysis of multi-environment trials
}
\examples{

library(metan)
library(dplyr)

# Based on stability only, for both GY and HM, higher is better
mtsi_model = waasb(data_ge,
                   env = ENV,
                   gen = GEN,
                   rep = REP,
                   resp = c(GY, HM))
mtsi_index = mtsi(mtsi_model)


# Based on mean performance and stability (using pipe operator \%>\%)
# GY: higher is better
# HM: lower is better
mtsi_index2 = data_ge \%>\%
              waasb(ENV, GEN, REP, c(GY, HM), mresp = c(100, 0)) \%>\%
              mtsi(index = "waasby")

}
\author{
Tiago Olivoto \email{tiagoolivoto@gmail.com}
}
