% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ammi.R
\name{performs_ammi}
\alias{performs_ammi}
\title{Additive Main effects and Multiplicative Interaction}
\usage{
performs_ammi(.data, env, gen, rep, resp, block = NULL, verbose = TRUE, ...)
}
\arguments{
\item{.data}{The dataset containing the columns related to Environments,
Genotypes, replication/block and response variable(s).}

\item{env}{The name of the column that contains the levels of the
environments}

\item{gen}{The name of the column that contains the levels of the genotypes}

\item{rep}{The name of the column that contains the levels of the
replications/blocks}

\item{resp}{The response variable(s). To analyze multiple variables in a
single procedure, use comma-separated list of unquoted variable names,
i.e., \code{resp = c(var1, var2, var3)}, or any select helper like
\code{resp = contains("_PLA")}.}

\item{block}{Defaults to \code{NULL}. In this case, a randomized complete
block design is considered. If block is informed, then a resolvable
alpha-lattice design (Patterson and Williams, 1976) is employed.
\strong{All effects, except the error, are assumed to be fixed.}}

\item{verbose}{Logical argument. If \code{verbose = FALSE} the code will run
silently.}

\item{...}{Arguments passed to the function
\code{\link[=impute_missing_val]{impute_missing_val()}} for imputation of missing values in case
of unbalanced data.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Compute the Additive Main effects and Multiplicative interaction (AMMI)
model.

This function is deprecated and will be removed in a future version.
Please use \code{\link[=ammi]{ammi()}} instead.
}
