% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plaisted_peterson.R
\name{plaisted_peterson}
\alias{plaisted_peterson}
\title{Stability analysis based on Plaisted and Peterson (1959)}
\usage{
plaisted_peterson(.data, env, gen, rep, resp, verbose = TRUE)
}
\arguments{
\item{.data}{The dataset containing the columns related to Environments,
Genotypes, replication/block and response variable(s).}

\item{env}{The name of the column that contains the levels of the
environments.}

\item{gen}{The name of the column that contains the levels of the genotypes.}

\item{rep}{The name of the column that contains the levels of the
replications/blocks.}

\item{resp}{The response variable(s). To analyze multiple variables in a
single procedure use, for example, \code{resp = c(var1, var2, var3)}.}

\item{verbose}{Logical argument. If \code{verbose = FALSE} the code will run
silently.}
}
\value{
An object of class \code{plaisted_peterson} containing the results for each
variable used in the argument \code{resp}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

The function computes phenotypic stability based on the method proposed by
Plaisted and Peterson (1959). It characterizes stability as the arithmetic mean
of the variance components of the genotype-environment interaction isolated
pairwise between a given genotype and all other remaining genotypes in the trial.
}
\details{
The Plaisted and Peterson method operates under a static (biological) concept of
stability. A genotype is classified as stable if its presence within the population
minimizes the total variance of the GxE interaction matrix.

For each pairwise combination of genotype \eqn{i} and genotype \eqn{i'}, the interaction
variance component \eqn{\sigma^2_{ga_{ii'}}} is partitioned. The overall stability index
(\eqn{\theta_i}) for a single cultivar is calculated as:

\deqn{\theta_i = \frac{1}{I - 1} \sum_{i' \neq i} \sigma^2_{ga_{ii'}}}

Where \eqn{I} is the total number of genotypes. The individual pairwise variance
components are computed by subtracting the pooled experimental error (\eqn{QMR})
from the pairwise interaction Mean Square:

\deqn{\sigma^2_{ga_{ii'}} = \frac{\left[ \frac{SS(g_{ii'} \times A)}{J - 1} \right] - QMR}{K}}

Where \eqn{J} is the total number of environments, \eqn{QMR} is the residual mean square
from the joint global ANOVA table, and \eqn{K} is the replication factor. The pairwise
interaction sum of squares is given by:

\deqn{SS(g_{ii'} \times A) = \frac{K}{2} \left[ d^2_{ii'} - \frac{1}{J}(Y_{i.} - Y_{i'.})^2 \right]}

Where \eqn{d^2_{ii'}} is the squared Euclidean distance between the performance profiles
of the two genotypes across environments:

\deqn{d^2_{ii'} = \sum_{j} (Y_{ij} - Y_{i'j})^2}

Cultivars displaying the lowest values of \eqn{\theta_i} contribute less to the population's
GxE noise and are interpreted as the most stable genotypes.
}
\examples{
\donttest{
library(metan)
 plaisted_peterson(data_ge, ENV, GEN, REP, GY)
}

}
\references{
Plaisted, R.L., and L.C. Peterson. 1959. A technique for
evaluating the ability of selections to yield consistently in different
locations or seasons. American Potato Journal 36(11): 381–385.
\doi{10.1007/BF02852735}
}
\author{
Tiago Olivoto \email{tiagoolivoto@gmail.com}
}
