% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlated_vars.R
\name{plot.correlated_vars}
\alias{plot.correlated_vars}
\title{Plot an object of class correlated_vars}
\usage{
\method{plot}{correlated_vars}(x, ...)
}
\arguments{
\item{x}{An object of class correlated_vars.}

\item{...}{Currently not used.}
}
\value{
An object of class gg.
}
\description{
Plot an object of class correlated_vars
}
\examples{
\donttest{
library(metan)
y <- rnorm(n = 10)
cor_vars <- correlated_vars(y, nvar = 6)
plot(cor_vars)
}
}
