% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.FAI_BLUP.R
\name{plot.FAI_BLUP}
\alias{plot.FAI_BLUP}
\title{Multitrait index based on factor analysis and ideotype-design}
\usage{
\method{plot}{FAI_BLUP}(x, ideotype = 1, SI = 15, radar = TRUE,
  size.point = 2, col.sel = "red", col.nonsel = "black", ...)
}
\arguments{
\item{x}{An object of class \code{WAASB}}

\item{ideotype}{The ideotype to be plotted. Default is 1.}

\item{SI}{An integer [0-100]. The selection intensity in percentage of the
total number of genotypes.}

\item{radar}{Logical argument. If true (default) a radar plot is generated
after using \code{coord_polar()}.}

\item{size.point}{The size of the point in graphic.}

\item{col.sel}{The colour for selected genotypes.}

\item{col.nonsel}{The colour for nonselected genotypes.}

\item{...}{Other arguments of the function.}
}
\description{
Plot the multitrait index based on factor analysis and ideotype-design
proposed by Rocha et al. (2018).
}
\examples{

library(metan)
library(dplyr)
multivariate = WAASB(data_ge,
                     resp = c(GY, HM),
                     gen = GEN,
                     env = ENV,
                     rep = REP)

FAI = data_ge2 \%>\%
      WAASB(ENV, GEN, REP, c(KW, NKE, PH, EH)) \%>\%
      FAI_BLUP(DI = c("max", "max", "max", "min"),
               UI = c("min", "min", "min", "max"),
               SI = 15)
      plot(FAI)

}
\references{
Rocha, J.R.A.S.C.R, J.C. Machado, and P.C.S. Carneiro. 2018.
Multitrait index based on factor analysis and ideotype-design: proposal and
application on elephant grass breeding for bioenergy. GCB Bioenergy
10:52-60. doi:
\href{https://onlinelibrary.wiley.com/doi/full/10.1111/gcbb.12443}{doi:10.1111/gcbb.12443}.
}
\author{
Tiago Olivoto \email{tiagoolivoto@gmail.com}
}
