% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ge_reg.R
\name{plot.ge_reg}
\alias{plot.ge_reg}
\title{Plot an object of class ge_reg}
\usage{
\method{plot}{ge_reg}(x, ...)
}
\arguments{
\item{x}{An object of class \code{ge_reg}.}

\item{...}{Other arguments of the function.}
}
\description{
Plot the regression model generated by the function \code{ge_reg}.
}
\author{
Tiago Olivoto \email{tiagoolivoto@gmail.com}
}
