% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.resp_surf.R
\name{plot.resp_surf}
\alias{plot.resp_surf}
\title{Plot the response surface model}
\usage{
\method{plot}{resp_surf}(x, xlab = NULL, ylab = NULL, region = TRUE,
  resolution = 100, ...)
}
\arguments{
\item{x}{An object of class \code{resp_surf}}

\item{xlab}{The label for the x axis}

\item{ylab}{The label for the y axis}

\item{region}{Logical argument indicating whether regions between contour lines
should be colored.}

\item{resolution}{The resolution of the contour plot. Defaults to 100.
higher values produce high-resolution plots but may increase the computation time.}

\item{...}{Other arguments passed from \code{contourplot} function.
See \code{?lattice::contourplot} for more details.}
}
\description{
Plot the response surface model using a contour plot
}
\author{
Tiago Olivoto \email{tiagoolivoto@gmail.com}
}
