% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_eigen.R
\name{plot_eigen}
\alias{plot_eigen}
\title{Plot the eigenvalues}
\usage{
plot_eigen(x, export = FALSE, theme = theme_waasb(),
  file.type = "pdf", file.name = NULL, width = 6, height = 6,
  size.shape = 3.5, size.line = 1, size.tex.lab = 12,
  y.lab = "Eigenvalue", y2.lab = "Accumulated variance",
  x.lab = "Number of multiplicative terms", resolution = 300, ...)
}
\arguments{
\item{x}{The \code{waasb object}}

\item{export}{Export (or not) the plot. Default is \code{TRUE}.}

\item{theme}{The graphical theme of the plot. Default is `theme =
theme_waasb()`. Please, see `?metan::theme_waasb`. An own theme can be
applied using the arguments: `theme = theme_waasb() + theme(some stuff
here)`. For more details, please, see `?ggplot2::theme`}

\item{file.type}{If \code{export = TRUE}, define the type of file to be
exported. Default is \code{pdf}, Graphic can also be exported in
\code{*.tiff} format by declaring \code{file.type = "tiff"}.}

\item{file.name}{The name of the file for exportation, default is
\code{NULL}, i.e. the files are automatically named.}

\item{width}{The width "inch" of the plot. Default is \code{6}.}

\item{height}{The height "inch" of the plot. Default is \code{6}.}

\item{size.shape}{The size of the shape. Default is \code{3.5}.}

\item{size.line}{The size of the line. Default is \code{1}.}

\item{size.tex.lab}{The size of the text in axis text and labels.}

\item{y.lab}{The label of the y-axis in the plot. Default is
\code{"Eigenvalue"}.}

\item{y2.lab}{The label of the second y-axis in the plot. Default is
\code{"Accumulated variance"}.}

\item{x.lab}{The label of the x-axis in the plot. Default is \code{"Number
of multiplicative terms"}.}

\item{resolution}{The resolution of the plot. Parameter valid if
\code{file.type = "tiff"} is used. Default is \code{300} (300 dpi)}

\item{...}{Other arguments of the function}
}
\value{
An object of class \code{gg, ggplot}.
}
\description{
Plot the eigenvalues for from singular value decomposition of BLUP
interaction effects matrix.
}
\examples{

library(metan)
BLUP = waasb(data_ge,
             resp = c(GY, HM),
             gen = GEN,
             env = ENV,
             rep = REP)
plot_eigen(BLUP$GY)



}
\seealso{
\code{\link{plot_scores}}, \code{\link{plot_waasby}}
}
\author{
Tiago Olivoto \email{tiagoolivoto@gmail.com}
}
