% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_factbars.R
\name{plot_factbars}
\alias{plot_factbars}
\title{Fast way to create a bar plot}
\usage{
plot_factbars(.data, ..., resp, y.expand = 1, y.breaks = waiver(),
  xlab = NULL, ylab = NULL, lab.bar = NULL, lab.bar.vjust = -0.2,
  size.text.bar = 5, errorbar = TRUE, stat.erbar = "se",
  width.erbar = 0.3, level = 0.95, invert = FALSE, col = TRUE,
  palette = "Spectral", width.bar = 0.9, lab.x.angle = 0,
  lab.x.hjust = 0.5, lab.x.vjust = 1, legend.position = "bottom",
  size.text = 12, fontfam = "sans", na.rm = TRUE, verbose = TRUE)
}
\arguments{
\item{...}{A comma-separated list of unquoted variable names. Must be up to two
variables.}

\item{resp}{The response variable}

\item{y.expand}{A multiplication factor to expand the y axis.. Defaults to 1.}

\item{y.breaks}{The breaks to be plotted in the y-axis. Defaults to waiver().
\code{authomatic breaks}. The same arguments than \code{x.breaks} can be
used.}

\item{xlab}{The x label}

\item{ylab}{The y label}

\item{lab.bar}{A vector of characters to show in each bar. Defaults to NULL.}

\item{lab.bar.vjust}{The vertical adjust for the labels in the bar. Defaults to -0.2}

\item{errorbar}{Logical argument, set to TRUE. In this case, an error bar is shown.}

\item{stat.erbar}{The statistic to be shown in the errorbar. Must be one of the 'se'
(standard error, default), 'sd' (standard deviation), or 'ci' confidence interval, based
on the confidence level}

\item{level}{The fonfidence level}

\item{invert}{Logical argument. If \code{TRUE}, the order of the factors entered
in changes in the graph}

\item{col}{Logical argument. If \code {FALSE}, a gray scale is used.}

\item{palette}{The color palette to be used. For more details, see
\code{?Scale_colour_brewer}}

\item{width.bar}{The width of the bars in the graph. Defaults to 0.9
possible values [0-1].}

\item{lab.x.angle}{The angle of the caption text. Default is 0.}

\item{lab.x.hjust}{The horizontal adjustment of the caption text. Defaults to 0.5.
Use this argument to adjust the text when the angle of the text is different from 0.}

\item{legend.position}{The legend position.}

\item{size.text}{The size of the text}

\item{fontfam}{The family of the font text}

\item{na.rm}{Should 'NA' values be removed to compute the statistics?
Defaults to true}

\item{verbose}{Logical argument. If FALSE the code will run silently}

\item{data}{The data set}

\item{alpha}{The alpha for the color in confidence band}

\item{size.shape}{The size for the shape in plot}

\item{size.line}{The size for the line in the plot}
}
\description{
Create a bar plot based on categorical (one or two) variables
and one numeric variable.
}
\seealso{
\code{\link{plot_lines}}, \code{\link{plot_factlines}}
}
