% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamem.R
\name{predict.gamem}
\alias{predict.gamem}
\title{Predict method for gamem fits}
\usage{
\method{predict}{gamem}(object, ...)
}
\arguments{
\item{object}{An object of class \code{gamem}}

\item{...}{Currently not used}
}
\value{
A tibble with the predicted values for each variable in the model
}
\description{
Obtains predictions from an object fitted with \code{\link{gamem}}.
}
\examples{
\donttest{
library(metan)
model <- gamem(data_g,
              gen = GEN,
              rep = REP,
              resp = everything())
predict(model)
}

}
\author{
Tiago Olivoto \email{tiagoolivoto@gmail.com}
}
