% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WAASB.R
\name{predict.waasb}
\alias{predict.waasb}
\title{Predict method for waasb fits}
\usage{
\method{predict}{waasb}(object, ...)
}
\arguments{
\item{object}{An object of class \code{waasb}}

\item{...}{Currently not used}
}
\value{
A tibble with the predicted values for each variable in the model
}
\description{
Obtains predictions from an object fitted with \code{\link[=waasb]{waasb()}}.
}
\examples{
\donttest{
library(metan)
model <- waasb(data_ge,
               env = ENV,
               gen = GEN,
               rep = REP,
               resp = c(GY, HM))
predict(model)
}

}
\author{
Tiago Olivoto \email{tiagoolivoto@gmail.com}
}
