% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_coef.R
\name{print.corr_coef}
\alias{print.corr_coef}
\title{Print an object of class corr_coef}
\usage{
\method{print}{corr_coef}(x, export = FALSE, file.name = NULL, digits = 3, ...)
}
\arguments{
\item{x}{An object of class \code{corr_coef}}

\item{export}{A logical argument. If \code{TRUE}, a *.txt file is exported to
the working directory.}

\item{file.name}{The name of the file if \code{export = TRUE}}

\item{digits}{The significant digits to be shown.}

\item{...}{Options used by the tibble package to format the output. See
\code{\link[tibble:formatting]{tibble::formatting()}} for more details.}
}
\description{
Print the \code{corr_coef} object in two ways. By default, the results
are shown in the R console. The results can also be exported to the directory
into a *.txt file.
}
\examples{
\donttest{
library(metan)
sel <- corr_coef(data_ge2, EP, EL, CD, CL)
print(sel)
}
}
\author{
Tiago Olivoto \email{tiagoolivoto@gmail.com}
}
