% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env_stratification.R
\name{print.env_stratification}
\alias{print.env_stratification}
\title{Print the env_stratification model}
\usage{
\method{print}{env_stratification}(x, export = FALSE, file.name = NULL, digits = 3, ...)
}
\arguments{
\item{x}{An object of class \code{env_stratification}.}

\item{export}{A logical argument. If \code{TRUE}, a *.txt file is exported to
the working directory}

\item{file.name}{The name of the file if \code{export = TRUE}}

\item{digits}{The significant digits to be shown.}

\item{...}{Currently not used.}
}
\description{
Print an object of class \code{ge_factanal} in two ways. By default, the results are
shown in the R console. The results can also be exported to the directory.
}
\examples{
\donttest{
library(metan)
model <-
env_stratification(data_ge,
                   env = ENV,
                   gen = GEN,
                   resp = GY)
print(model)
}
}
\author{
Tiago Olivoto \email{tiagoolivoto@gmail.com}
}
