% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.ge_factanal.R
\name{print.ge_factanal}
\alias{print.ge_factanal}
\title{Print an object of class ge_factanal}
\usage{
\method{print}{ge_factanal}(x, export = FALSE, file.name = NULL, digits = 4, ...)
}
\arguments{
\item{x}{An object of class \code{ge_factanal}.}

\item{export}{A logical argument. If \code{TRUE}, a *.txt file is exported to
the working directory}

\item{file.name}{The name of the file if \code{export = TRUE}}

\item{digits}{The significant digits to be shown.}

\item{...}{Options used by the tibble package to format the output. See
\code{\link[tibble]{trunc_mat}} for more details.}
}
\description{
Print the \code{ge_factanal} object in two ways. By default, the results are
shown in the R console. The results can also be exported to the directory.
}
\examples{

model <- ge_factanal(data_ge2,
  env = ENV,
  gen = GEN,
  rep = REP,
  resp = PH
)
print(model)
}
\author{
Tiago Olivoto \email{tiagoolivoto@gmail.com}
}
