% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ammi.R
\name{print.performs_ammi}
\alias{print.performs_ammi}
\title{Print an object of class performs_ammi}
\usage{
\method{print}{performs_ammi}(x, export = FALSE, file.name = NULL, digits = 4, ...)
}
\arguments{
\item{x}{An object of class \code{performs_ammi}.}

\item{export}{A logical argument. If \code{TRUE}, a *.txt file is exported to
the working directory}

\item{file.name}{The name of the file if \code{export = TRUE}}

\item{digits}{The significant digits to be shown.}

\item{...}{Options used by the tibble package to format the output. See
\code{\link[tibble:formatting]{tibble::print()}} for more details.}
}
\description{
Print the \code{performs_ammi} object in two ways. By default, the results are shown
in the R console. The results can also be exported to the directory.
}
\examples{
\donttest{
library(metan)
model <- ammi(data_ge, ENV, GEN, REP,
                       resp = c(GY, HM))
print(model)
}
}
\author{
Tiago Olivoto \email{tiagoolivoto@gmail.com}
}
