% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.can_cor.R
\name{summary.can_cor}
\alias{summary.can_cor}
\title{Summarise an object of class can_cor}
\usage{
\method{summary}{can_cor}(object, export = FALSE, file.name = NULL,
  digits = 3, ...)
}
\arguments{
\item{object}{The \code{can_cor} object}

\item{export}{A logical argument. If \code{TRUE|T}, a *.txt file is exported
to the working directory}

\item{file.name}{The name of the file if \code{export = TRUE}}

\item{digits}{The significant digits to be shown.}

\item{...}{Other arguments of the function}
}
\description{
Summarise an object of class \code{can_cor} object in two ways. By default,
the results are shown in the R console. The results can also be exported to
the directory.
}
\examples{

\dontrun{
library(METAAB)
cc = can_corr(data_ge2,
              FG = c(PH, EH, EP),
              SG = c(EL, CL, CD, CW, KW, NR, TKW),
              verbose = FALSE)
summary(cc)
summary(cc,
        export = TRUE,
        file.name = "canonical results")
}

}
\author{
Tiago Olivoto \email{tiagoolivoto@gmail.com}
}
