% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.ge_reg.R
\name{summary.ge_reg}
\alias{summary.ge_reg}
\title{Summary a ge_reg object}
\usage{
\method{summary}{ge_reg}(object, export = FALSE, file.name = NULL,
  digits = 3, ...)
}
\arguments{
\item{object}{The \code{ge_reg} object}

\item{export}{A logical argument. If \code{TRUE}, a *.txt file is exported
to the working directory.}

\item{file.name}{The name of the file if \code{export = TRUE}}

\item{digits}{The significant digits to be shown.}

\item{...}{Other arguments of the function}
}
\description{
Summary the \code{ge_reg} object in two ways. By default, the results
are shown in the R console. The results can also be exported to the
directory into a *.txt file.
}
\examples{

library(METAAB)
model = ge_reg(data_ge2, ENV, GEN, REP, PH)
summary(model)

}
\author{
Tiago Olivoto \email{tiagoolivoto@gmail.com}
}
