% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.ge_stats.R
\name{summary.ge_stats}
\alias{summary.ge_stats}
\title{Summary a ge_stats object}
\usage{
\method{summary}{ge_stats}(object, export = FALSE, file.name = NULL,
  digits = 3, ...)
}
\arguments{
\item{object}{The \code{ge_stats} object}

\item{export}{A logical argument. If \code{TRUE}, a *.txt file is exported
to the working directory.}

\item{file.name}{The name of the file if \code{export = TRUE}}

\item{digits}{The significant digits to be shown.}

\item{...}{Other arguments of the function}
}
\description{
Summary the \code{ge_stats} object in two ways. By default, the results are
shown in the R console. The results can also be exported to the directory
into a *.txt file.
}
\examples{

library(METAAB)
model = ge_stats(.data = data_ge2,
                 env = ENV,
                 gen = GEN,
                 rep = REP,
                 resp = PH)
summary(model)

}
\author{
Tiago Olivoto \email{tiagoolivoto@gmail.com}
}
