% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.WAASB.R
\name{summary.WAASB}
\alias{summary.WAASB}
\title{Summary a WAASB object}
\usage{
\method{summary}{WAASB}(object, export = FALSE, blup = FALSE,
  file.name = NULL, digits = 4, ...)
}
\arguments{
\item{object}{The \code{WAASB} object}

\item{export}{A logical argument. If \code{TRUE|T}, a *.txt file is exported
to the working directory}

\item{blup}{A logical argument. If \code{TRUE|T}, the blups are shown.}

\item{file.name}{The name of the file if \code{export = TRUE}}

\item{digits}{The significant digits to be shown.}

\item{...}{Other arguments of the function}
}
\description{
Summary the \code{WAASB} object in two ways. By default, the results are
shown in the R console. The results can also be exported to the directory.
}
\examples{

library(METAAB)
model = WAASB(data_ge,
              resp = c(GY, HM),
              gen = GEN,
              env = ENV,
              rep = REP)
summary(model)

}
\author{
Tiago Olivoto \email{tiagoolivoto@gmail.com}
}
