% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_binns.R
\name{superiority}
\alias{superiority}
\title{Lin e Binns' superiority index (Legacy Alias)}
\usage{
superiority(.data, env, gen, resp, verbose = TRUE)
}
\arguments{
\item{.data}{The dataset containing the columns related to Environments,
Genotypes, replication/block and response variable(s)}

\item{env}{The name of the column that contains the levels of the
environments.}

\item{gen}{The name of the column that contains the levels of the genotypes.}

\item{resp}{The response variable(s). To analyze multiple variables in a
single procedure use, for example, \code{resp = c(var1, var2, var3)}.}

\item{verbose}{Logical argument. If \code{verbose = FALSE} the code will run
silently.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{superiority()} was deprecated to match cleaner naming conventions.
Please transition script targets over to using \code{lin_binns()}.
}
\keyword{internal}
