\name{theme_waasb}
\alias{theme_waasb}
\title{A personalized theme for the WAASB ggplot2-based graphics}

\description{
The function \code{theme_waasb()} is a personalized theme for the WAASB ggplot2-based graphics. An own theme can be applied in all graphics by using the following arguments in the \code{plot.} functions of the package: \code{theme = theme_waasb() + theme(some stuff here...)}. See \code{?ggplot2::theme} for more details.}


\author{Tiago Olivoto \email{tiagoolivoto@gmail.com}}
