% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{utilities}
\alias{utilities}
\alias{round_value}
\alias{round_column}
\alias{column_exists}
\title{Utilities for manipulating data}
\usage{
round_value(x, digits = 2)

round_column(.data, ..., digits = 2)

column_exists(.data, cols)
}
\arguments{
\item{x}{A numeric value or a numeric vector.}

\item{digits}{The number of significant figures.}

\item{.data}{A data frame}

\item{...}{Variables to round. If no variable is informed, all the numeric
variables from \code{data} are used.}

\item{cols}{A quoted variable name.}
}
\description{
Utilities for manipulating data.
\itemize{
\item \code{round_value()}: Round a numeric value or a numeric vector to significant figures.
\item \code{round_column()}: Round a selected column or a whole data frame to significant figures.
\item \code{column_exists()}: Check if a column exists in a data frame.
}
}
