% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WAAS.AMMI.R, R/predict.WAAS.AMMI.R,
%   R/summary.WAAS.AMMI.R
\name{WAAS.AMMI-deprecated}
\alias{WAAS.AMMI-deprecated}
\title{Weighted Average of Absolute Scores}
\usage{
WAAS.AMMI(.data, env, gen, rep, resp, mresp = NULL, wresp = NULL,
  prob = 0.05, naxis = NULL, verbose = TRUE)

WAAS.AMMI(.data, env, gen, rep, resp, mresp = NULL, wresp = NULL,
  prob = 0.05, naxis = NULL, verbose = TRUE)

WAAS.AMMI(.data, env, gen, rep, resp, mresp = NULL, wresp = NULL,
  prob = 0.05, naxis = NULL, verbose = TRUE)
}
\arguments{
\item{.data}{The dataset containing the columns related to Environments,
Genotypes, replication/block and response variable(s).}

\item{env}{The name of the column that contains the levels of the
environments.}

\item{gen}{The name of the column that contains the levels of the genotypes.}

\item{rep}{The name of the column that contains the levels of the
replications/blocks.}

\item{resp}{The response variable(s). To analyze multiple variables in a
single procedure a vector of variables may be used. For example \code{resp =
c(var1, var2, var3)}.}

\item{mresp}{A numeric vector of the same length of \code{resp}. The
\code{mresp} will be the new maximum value after reescaling. By default, all
variables in \code{resp} are rescaled so that de maximum value is 100 and
the minimum value is 0.}

\item{wresp}{The weight for the response variable(s) for computing the
WAASBY index. Must be a numeric vector of the same length of \code{resp}.
Defatul is 50, i.e., equal weights for stability and mean performance.}

\item{prob}{The p-value for considering a IPCA significant.}

\item{naxis}{The number of IPCAs to be used for computing the WAAS index.
Default is \code{NULL} (Significant IPCAs are used). If values are informed,
the number of IPCAS will be used independently on its significance. Note
that if two or more variables are included in \code{resp}, then \code{naxis}
must be a vector.}

\item{verbose}{Logical argument. If \code{verbose = FALSE} the code is run
silently.}

\item{object}{An object of class WAAS.AMMI}

\item{naxis}{The the number of axis to be use in the prediction. If
\code{object} has more than one variable, then \code{naxis} must be a
vector.}

\item{...}{Additional parameter for the function}

\item{object}{The \code{WAAS.AMMI} object}

\item{export}{A logical argument. If \code{TRUE|T}, a *.txt file is exported
to the working directory}

\item{file.name}{The name of the file if \code{export = TRUE}}

\item{digits}{The significant digits to be shown.}

\item{...}{Other arguments of the function}
}
\description{
Weighted Average of Absolute Scores

Predict the means of a WAAS.AMMI object considering a specific number of
axis.

Summary the \code{WAAS.AMMI} object in two ways. By default, the results are
shown in the R console. The results can also be exported to the directory.
}
\details{
This function is used to predict the response variable of a two-way table
(for examples the yielding of the i-th genotype in the j-th environment)
based on AMMI model. This prediction is based on the number of
multiplicative terms used. If \code{naxis = 0}, only the main effects
(AMMI0) are used. In this case, the predicted mean will be the predicted
value from OLS estimation. If \code{naxis = 1} the AMMI1 (with one
multiplicative term) is used for predicting the response variable. If
\code{naxis = min(gen-1;env-1)}, the AMMIF is fitted and the predicted value
will be the cell mean, i.e. the mean of R-replicates of the i-th genotype in
the j-th environment. The number of axis to be used must be carrefully
chosen. Precures based on Postdictive sucess (such as Gollobs's d.f.) or
Predictive sucess (such as cross-validation) should be used to do this. This
package provide both. \code{\link{WAAS.AMMI}} function compute traditional
AMMI analysis showing the number of significant axis. On the other hand,
\code{\link{cv_ammif}} function provide a cross-validation,
estimating the RMSPD of all AMMI-family models, based on resampling
procedures.
}
\seealso{
\code{\link{metan-deprecated}}

\code{\link{metan-deprecated}}

\code{\link{metan-deprecated}}
}
\author{
Tiago Olivoto \email{tiagoolivoto@gmail.com}

Tiago Olivoto \email{tiagoolivoto@gmail.com}

Tiago Olivoto \email{tiagoolivoto@gmail.com}
}
\keyword{internal}
