\name{WAASBdata}
\alias{WAASBdata}
\title{A replicate-based data of 10 genotypes cultivated in 11 environments}

\description{

 This dataset contain the observed grain yield of 10 genotypes cultivated in 11 environments. The experimental design was a RCBD with 3 replicates(blocks). This data provide examples for several functions of \code{WAAS} package.
}
\author{Tiago Olivoto \email{tiagoolivoto@gmail.com}}

