% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WAASBYratio.R, R/WAASratio.AMMI.R,
%   R/plot.WAASBYratio.R
\name{WAASBYratio-deprecated}
\alias{WAASBYratio-deprecated}
\title{Different scenarios of stability and mean performance}
\usage{
WAASBYratio(.data, env, gen, rep, resp, mresp = 100, increment = 10,
  saveWAASY = 50, progbar = TRUE)

WAASBYratio(.data, env, gen, rep, resp, mresp = 100, increment = 10,
  saveWAASY = 50, progbar = TRUE)

WAASBYratio(.data, env, gen, rep, resp, mresp = 100, increment = 10,
  saveWAASY = 50, progbar = TRUE)
}
\arguments{
\item{.data}{The dataset containing the columns related to Environments,
Genotypes, replication/block and response variable(s).}

\item{env}{The name of the column that contains the levels of the
environments.}

\item{gen}{The name of the column that contains the levels of the genotypes.}

\item{rep}{The name of the column that contains the levels of the
replications/blocks.}

\item{resp}{The response variable.}

\item{mresp}{A numeric value that will be the new maximum value after reescaling.
By default, the variable in \code{resp} is rescaled so that the original maximum
 and minimum values are 100 and 0, respectively.}

\item{increment}{The range of the increment for WAASB/GY ratio. Default is
5. The function compute the WAASBY values starting with a weight o 100 for
stability and 0 for response variable. With the default, the first scenario
will be a WAASB/GY ratio = 100/0. In the next scenario, the WAASBY values
are computed based on a WAASB/GY ratio = 95/5.}

\item{saveWAASY}{Automatically save the WAASY values when the wheight for
WAAS (stability) in the WAAS/GY ratio is "saveWAASY". Default is 100. The
value of "saveWAASY" must be multiple of "Increment". If this assumption is
not valid, an error will be occour.}

\item{progbar}{A logical argument to define if a progress bar is shown.
Default is \code{TRUE}.}

\item{.data}{The dataset containing the columns related to Environments,
Genotypes, replication/block and response variable(s).}

\item{env}{The name of the column that contains the levels of the
environments.}

\item{gen}{The name of the column that contains the levels of the genotypes.}

\item{rep}{The name of the column that contains the levels of the
replications/blocks.}

\item{resp}{The response variable.}

\item{mresp}{A numeric value that will be the new maximum value after reescaling.
By default, the variable in \code{resp} is rescaled so that the original maximum
 and minimum values are 100 and 0, respectively.}

\item{p.valuePC}{The p-value for considering the PC significant. Default is
0.05. The number of significant Principal Components to be used for
calculating the WAASB will be chosen based on this probability.}

\item{increment}{The range of the increment for WAAS/GY ratio. Default is 5.
The function compute the WAASY values starting with a weight o 100 for
stability and 0 for response variable. With the default, the first scenario
will be a WAAS/GY ratio = 100/0. In the next scenario, the WAASY values are
computed based on a WAAS/GY ratio = 95/5.}

\item{saveWAASY}{Automatically save the WAASY values when the wheight for
WAAS (stability) in the WAAS/GY ratio is "saveWAASY". Default is 100. The
value of "saveWAASY" must be multiple of "Increment". If this assumption is
not valid, an error will be occour.}

\item{progbar}{A logical argument to define if a progress bar is shown.
Default is \code{TRUE}.}

\item{x}{The \code{WAASBYratio object}}

\item{type}{\code{1 = Heat map Ranks}: this graphic shows the genotype
ranking considering the WAAS estimated with different numbers of Principal
Components; \code{2 = Heat map WAASY-GY ratio}: this graphic shows the
genotype ranking considering the different combinations in the WAAS/GY
ratio.}

\item{export}{Export (or not) the plot. Default is \code{TRUE}.}

\item{file.type}{If \code{export = TRUE} define the type of file to be
exported. Default is \code{pdf}, Graphic can also be exported in
\code{*.tiff} format by declaring \code{file.type = "tiff"}.}

\item{file.name}{The name of the file for exportation, default is
\code{NULL}, i.e. the files are automatically named.}

\item{width}{The width "inch" of the plot. Default is \code{8}.}

\item{height}{The height "inch" of the plot. Default is \code{7}.}

\item{size.lab}{The label size of the plot. It is suggested attribute 1}

\item{margins}{Numeric vector of length 2 containing the margins for column
and row names, respectively. Default is \code{c(5, 4)}.}

\item{y.lab}{The label of y axis. Default is "Genotypes".}

\item{x.lab}{The label of x axis. Default is "Number of axes".}

\item{key.lab}{The label of color key. Default is "Genotype ranking".}

\item{resolution}{Valid parameter if \code{file.type = "tiff"}. Define the
resolution of the plot. Default is "300".}

\item{...}{Other arguments of the function}
}
\value{
\item{anova}{Joint analysis of variance for the main effects and
Principal Component analysis of the interaction effect.}

\item{PC}{Principal Component Analysis.}

\item{MeansGxE}{The means of genotypes in the environments, with observed,
predicted and residual values.}

\item{WAAS}{A data frame with the response variable, the scores of all
Principal Components, the estimates of Weighted Average of Absolute Scores,
and WAASY (the index that consider the weights for stability and
productivity in the genotype ranking.}

\item{WAASY}{The values of the WAASY estimated when the wheight for the
stability in the loop match with argument "saveWAASY".}

\item{WAASY.values}{All the values of WAASY estimated in the different
scenarios of WAAS/GY weighting ratio.}
}
\description{
This function computes the WAASBY index in mixed-effec model analysis in
different combinations of weights for stability and productivity.

This function computes the WAASBY index in AMMI analysis in
different combinations of weights for stability and productivity.

Plot heat maps with genotype ranking in two ways.
}
\details{
This function considers both stability (weighted average of absolute scores
based on SVD of BLUP-interaction effects matrix) and productivitye for
genotype ranking. This function provide the option of attributing weights
for stability and productive in genotype ranking. This is important
depending on the goal of a selection strategy. For example, if a a goal of a
breeding program is to select a genotype whith high yielding (independeltly
on the stability performance), that genotype with the first rank in an
WAASB/GY = 0/100 ratio should be selected. The reciprocal is true. Aiming at
selecting a high-stable genotype (independentely on the productivity), that
genotype with the first rank in an WAASB/GY = 100/0 ratio should be
selected. By defalut, the increment on the WAASB/GY ratio is equal to 5. In
other words, twenty one different combinations are computed. Each
combination, the genotypes are ranked regarding the WAASY value.

This function is very similar to the \code{WAASBYratio}. The main difference
is that here, the WAASBY values are computed considering a traditional AMMI
model.

The first type of heatmap shows the genotype ranking depending on the number
of principal component axis used for estimating the WAASB index. An
euclidian distance-based dendrogram is used for grouping the genotype
ranking for both genotypes and principal component axis. The second type of
heatmap shows the genotype ranking depending on the WAASB/GY ratio. The
ranks obtained with a ratio of 100/0 considers exclusively the stability for
the genotype ranking. On the other hand, a ratio of 0/100 considers
exclusively the productivity for the genotype ranking.  Four clusters are
estimated (1) unproductive and unstable genotypes; (2) productive, but
unstable genotypes; (3) stable, but unproductive genotypes; and (4),
productive and stable genotypes.
}
\seealso{
\code{\link{metan-deprecated}}

\code{\link{metan-deprecated}}

\code{\link{metan-deprecated}}
}
\author{
Tiago Olivoto \email{tiagoolivoto@gmail.com}

Tiago Olivoto \email{tiagoolivoto@gmail.com}
}
\keyword{internal}
