% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colindiag.R
\name{colindiag}
\alias{colindiag}
\title{Collinearity Diagnostics}
\usage{
colindiag(.data, n = NULL, verbose = TRUE)
}
\arguments{
\item{.data}{The data to be analyzed. Must be a symmetric correlation matrix
or, a dataframe containing the predictor variables, or an object of class
\code{split_factors}.}

\item{n}{If a correlation matrix is provided, then \code{n} is the number of
objects used to compute the correlation coefficients.}

\item{verbose}{If \code{verbose = TRUE} then some results are shown in the
console.}
}
\value{
The following values are returned. Please, note that if a grouping variable
is used, then the results are returned into a list.

\item{cormat}{A symmetric Pearson's coefficient correlation matrix between
the variables}

\item{corlist}{A hypothesis testing for each of the correlation
coefficients}

\item{evalevet}{The eigenvalues with associated eigenvectors of the
correlation matrix}

\item{VIF}{The Variance Inflation Factors, being the diagonal elements of
the inverse of the correlation matrix.}

\item{CN}{The Condition Number of the correlation matrix, given by the ratio
between the largest and smallest eigenvalue.}

\item{det}{The determinant of the correlation matrix.}

\item{largest_corr}{The largest correlation (in absolute value) observed.}

\item{smallest_corr}{The smallest correlation (in absolute value) observed.}

\item{weight_var}{The variables with largest eigenvector (largest weight) in
the eigenvalue of smallest value, sorted in decreasing order.}
}
\description{
Perform a (multi)collinearity diagnostic of a correlation matrix of
predictor variables based several indicators, as shown by Olivoto et al.
(2017).
}
\examples{

# Using the correlation matrix
library(metan)
cor_iris = cor(iris[,1:4])
n = nrow(iris)
colindiag(cor_iris, n = n, verbose = FALSE)

# Using the pipe operator \%>\%
cor_iris \%>\% colindiag(n = n, verbose = FALSE)

# Diagnostic by species, storing into an object
col_diag_spec = iris \%>\%
                split_factors(Species) \%>\%
                colindiag()


}
\references{
Olivoto, T., V.Q. Souza, M. Nardino, I.R. Carvalho, M. Ferrari,
A.J. Pelegrin, V.J. Szareski, and D. Schmidt. 2017. Multicollinearity in
path analysis: a simple method to reduce its effects. Agron. J. 109:131-142.
doi:10.2134/agronj2016.04.0196.
\href{https://dl.sciencesocieties.org/publications/aj/abstracts/109/1/131}{doi:10.2134/agronj2016.04.0196}
}
\author{
Tiago Olivoto \email{tiagoolivoto@gmail.com}
}
