% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{data_ge}
\alias{data_ge}
\title{A replicate-based data of 10 genotypes cultivated in 14 environments}
\arguments{
\item{ENV}{A factor with 14 levels; each level represents one cultivation
environment.}

\item{GEN}{A factor with 10 levels; each level represents one genotype.}

\item{REP}{A factor with 3 levels; each level represents one
replication/block.}

\item{GY}{A continuous variable (grain yield) observed in each plot.}

\item{HM}{A continuous variable (hectoliter mass) observed in each plot.}
}
\description{
This dataset contain data on two variables assessed in 10 genotypes growing
in in 11 environments. The experimental design was a RCBD with 3
replicates(blocks). This data provide examples for several functions of
\pkg{metan} package.
}
\author{
Tiago Olivoto \email{tiagoolivoto@gmail.com}
}
\keyword{data}
