% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{data_ge2}
\alias{data_ge2}
\title{A replicate-based data of 13 maize hybrids cultivated in 4 environments}
\arguments{
\item{ENV}{A factor with 4 levels; each level represents one cultivation
environment.}

\item{GEN}{A factor with 13 levels; each level represents one maize hybrid.}

\item{REP}{A factor with 3 levels; each level represents one
replication/block.}

\item{PH}{Plant height, in cm.}

\item{EH}{Ear height, in cm.}

\item{EP}{Ear position, i.e., the ratio EH/PH.}

\item{EL}{Ear length, in cm.}

\item{ED}{Ear diameter, in mm.}

\item{CL}{Cob length, in cm.}

\item{CD}{Cob diameter, in mm.}

\item{CW}{Cob weight, in g.}

\item{KW}{Kernel weight, in cm.}

\item{NR}{Number of rows.}

\item{NKR}{Number of kernels per row.}

\item{CDED}{Cob diameter / Ear diameter ratio.}

\item{PERK}{Percentage of kernels.}

\item{TKW}{Thousand-kernel weight}

\item{NKE}{Number of kernels per row.}
}
\description{
This dataset contain data on 15 traits assessed in 13 maize hybrids growing
in 4 environments. The experimental design was a RCBD with 3 blocks and 1
replications per block. It may be used as example in several functions of
\pkg{METAAB} package.
}
\author{
Tiago Olivoto \email{tiagoolivoto@gmail.com}
}
\keyword{data}
