\name{datameans}
\alias{datameans}
\title{A dataset with means of 10 genotypes cultivated in 5 environments }

\description{
This dataset contains the means for grain yield of 10 genotypes cultivated in 5 environments. The interaction effects for this data is found in \link{intmatrix}

}
<<<<<<< HEAD

=======
>>>>>>> 071f2768c99ac8d77f2e71d73ed3ab15ac135a6a
\author{Tiago Olivoto \email{tiagoolivoto@gmail.com}}
