% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_stat.R
\name{desc_stat}
\alias{desc_stat}
\title{Descriptive statistics}
\usage{
desc_stat(.data = NULL, ..., values = NULL, stats = NULL,
  level = 0.95, digits = 4, verbose = TRUE)
}
\arguments{
\item{.data}{The data to be analyzed. Must be a dataframe or an object of
class \code{split_factors}.}

\item{...}{A single variable name or a comma-separated list of unquoted variables names.}

\item{values}{An alternative way to pass the data to the function. It must be a numeric
vector.}

\item{stats}{The descriptive statistics to show. Defaults to \code{NULL} (all statistics shown).
 It must be one of the \code{'AV.dev'} (average deviation),
\code{'CI.mean'} (confidence interval for the mean), \code{'CV'} (coefficient of variation),
\code{'IQR'} (interquartile range), \code{'Kurt'} (kurtosis), \code{'max'} (maximum value),
\code{'mean'} (arithmetic mean), \code{'median'} (median), \code{'min'} (minimum value),
\code{'n'} (the length of the data), \code{'norm.pval'} (the p-value for the Shapiro-Wilk test),
 \code{'norm.stat'} (the statistic for the Shapiro-Wilk test), \code{'Q2.5'} (the percentile 2.5%),
 \code{'Q25'} (the first quartile, Q1), \code{'Q75'} (the third quartile, Q3), \code{'Q97.5'} (the percentile 97.5%),
 \code{'SD.amo'} (the sample standard deviation), \code{'SD.pop'} (the population standard deviation),
 \code{'SE.mean'} (the standard error of the mean), \code{'skew'} (the skewness), \code{'var.amo'} (the sample variance),
 \code{'var.pop'} (the population variance). Use a comma-separated vector of names to select the statistics.
 For example, \code{stats = c("media, mean, CV, n")}.}

\item{level}{The confidence level to compute the confidence interval of mean. Defaults to 0.95.}

\item{digits}{The number of significant figures.}

\item{verbose}{Logical argument. If \code{verbose = FALSE} the code is run silently.}
}
\value{
A tibble with the statistics in the lines and variables in columns. If
\code{.data} is an object of class \code{split_factors}, then the statistics will be
shown for each level of the grouping variable in the function \code{split_factors()}.
}
\description{
Compute the most used measures of central tendency, position, and dispersion.
}
\examples{
library(metan)
library(dplyr)
desc1 <- data_ge2 \%>\% desc_stat(TKW)

vect <- data_ge2 \%>\%
  select(TKW) \%>\%
  pull()

desc2 <- desc_stat(values = vect)

desc3 <-
data_ge2 \%>\%
  split_factors(ENV) \%>\%
  desc_stat(EP, EL, PH, CL, CW, NR, NKR,
  stats = c('mean, SE.mean, CV'),
  verbose = FALSE)

}
\author{
Tiago Olivoto \email{tiagoolivoto@gmail.com}
}
