% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_data.R
\name{get_model_data}
\alias{get_model_data}
\title{Easily get data from a model}
\usage{
get_model_data(x, what = "Y", type = "GEN")
}
\arguments{
\item{x}{An object created with the functions \code{\link{waas}} or
\code{\link{waasb}}.}

\item{what}{What should be captured from the model. See more in
\strong{Details} section.}

\item{type}{Chose if the statistics must be show by genotype (\code{type =
"GEN"}, default) or environment (\code{type = "ENV"}), when possible.}
}
\value{
A tibble showing the values of the variable chosen in argument
\code{what}.
}
\description{
Easily get data from a model of class \code{waas} or \code{waasb} (Olivoto et
al., 2019a, 2019b)
}
\details{
\strong{The next options are valid in the argument \code{what} for both \code{waas} and \code{waasb} objects.}
\itemize{
\item \code{"Y"} for raw means (Default) .
\item \code{"PCn"} for the scores of the nth Interaction Principal Component Axis (IPCA) .
where * \code{n} is the desired IPCA (e.g. \code{"PC3"}) .
\item \code{"WAASB"}  The weighted average of the absolute scores .
\item \code{"PctResp"} The rescaled values of the response variable .
\item \code{"PctWAASB"} The rescaled values of the WAASB .
\item \code{"wResp"} The weight for the response variable .
\item \code{"wWAASB"} The weight for the stability .
\item \code{"OrResp"} The ranking regarding the response variable .
\item \code{"OrWAASB"} The ranking regarding the WAASB .
\item \code{"OrPC1"} The ranking regarding the first principal component axix .
\item \code{"WAASBY"} The superiority index WAASBY .
\item \code{"OrWAASBY"} The ranking regarding the superiority index .
}

\strong{The next options are only allowed if the object is of class \code{waasb}.}
\itemize{
\item \code{"blupg"} For genotype's predicted mean.
\item \code{"blupge"} for genotype-vs-environment's predicted mean.
\item \code{"genpar"} Genetic parameters.
\item \code{"lrt"} The statistic for the likelihood-ratio test for random effects.
\item \code{"pval_lrt"}  The p-values for the likelihood-ratio test.
\item \code{"vcomp"} The variance components for random effects.
}
}
\examples{

library(metan)
# Fit an AMMI model for 5 variables.
AMMI <- waas(data_ge2, ENV, GEN, REP,
             resp = c(PH, ED, TKW, NKR))

# Getting the weighted average of absolute scores
get_model_data(AMMI, what = "WAASB")

# And the rank for the WAASB inde.
get_model_data(AMMI, what = "OrWAASB")

# Fitting a mixed-effect model
blup <- waasb(data_ge2, ENV, GEN, REP,
              resp = c(PH, ED, TKW, NKR))

# Getting p-values for likelihood-ratio test
get_model_data(blup, what = "pval_lrt")

# Getting the variance components
get_model_data(blup, what = "vcomp")

# Getting the genetic parameters
get_model_data(blup, what = "genpar")


}
\references{
Olivoto, T., A.D.C. L{'{u}}cio, J.A.G. da silva, V.S. Marchioro,
V.Q. de Souza, and E. Jost. 2019a. Mean performance and stability in multi-environment
trials I: Combining features of AMMI and BLUP techniques. Agron. J.
\href{https://dl.sciencesocieties.org/publications/aj/abstracts/0/0/agronj2019.03.0220?access=0&view=pdf}{doi:10.2134/agronj2019.03.0220}

Olivoto, T., A.D.C. L{'{u}}cio, J.A.G. da silva, B.G. Sari, and M.I. Diel. 2019b.
Mean performance and stability in multi-environment trials II: Selection based on multiple
traits. Agron. J. (in press)
}
\author{
Tiago Olivoto \email{tiagoolivoto@gmail.com}
}
