% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_data.R
\name{get_model_data}
\alias{get_model_data}
\title{Easily get data from a model}
\usage{
get_model_data(x, what = "Y", type = "GEN")
}
\arguments{
\item{x}{An object created with the functions \code{\link{AMMI_indexes}},
\code{\link{ecovalence}},  \code{\link{Fox}},  \code{\link{ge_reg}},
\code{\link{gamem}}, \code{\link{performs_ammi}},
\code{\link{Resende_indexes}}, \code{\link{Shukla}},
\code{\link{superiority}}, \code{\link{waas}} or \code{\link{waasb}}.}

\item{what}{What should be captured from the model. See more in
\strong{Details} section.}

\item{type}{Chose if the statistics must be show by genotype (\code{type =
"GEN"}, default) or environment (\code{type = "ENV"}), when possible.}
}
\value{
A tibble showing the values of the variable chosen in argument
\code{what}.
}
\description{
Easily get data from some objects generated in the \strong{metan} package
such as the WAASB and WAASBY indexes  (Olivoto et al., 2019a, 2019b) BLUPs,
variance components, details of AMMI models and AMMI-based stability
statistics
}
\details{
Bellow are listed the options allowed in the argument \code{what} depending
on the class of the object

\strong{Object is of class \code{AMMI_indexes}.}
\itemize{
\item \code{"Y"} for raw means (Default).
\item \code{"ASV"} AMMI stability value
\item \code{"SIPC"} Sums of the absolute value of the IPCA scores
\item \code{"EV"} Averages of the squared eigenvector values
\item \code{"ZA"} Absolute value of the relative contribution of IPCAs to the
interaction
\item \code{"WAAS"} P-value for for each IPCA.
}

\strong{Object is of class \code{ecovalence}.}
\itemize{
\item \code{"Ecoval"} Ecovalence value.
\item \code{"Ecov_perc"} Ecovalence in percentage value.
\item \code{"rank"} Rank for ecovalence
}

\strong{Object is of class \code{ge_reg}.}
\itemize{
\item \code{"Y"} for raw means (Default).
\item \code{"slope"} The sloop of the regression
\item \code{"deviations"} The deviations from regression
\item \code{"RMSE"} The Root Mean Square Error
\item \code{"R2"} The r-square of the regression
}

\strong{Object is of class \code{Fox}.}
\itemize{
\item \code{"Y"} for raw means (Default).
\item \code{"TOP"} The proportion of locations at which the genotype occurred in
the top third
}

\strong{Object is of class \code{Shukla}.}
\itemize{
\item \code{"Y"} for raw means (Default).
\item \code{"rMean"} Rank for the mean.
\item \code{"ShuklaVar"} Shukla's stablity variance.
\item \code{"rShukaVar"} Rank for Shukla's stablity variance.
\item \code{"ssiShukaVar"} Simultaneous selection index.

\strong{Object is of class \code{superiority}.}
\item \code{"Y"} for raw means (Default).
\item \code{"Pi_a"} The superiority measure for all environments.
\item \code{"R_a"} The rank for Pi_a.
\item \code{"Pi_f"} The superiority measure for favorable environments.
\item \code{"R_f"} The rank for Pi_f.
\item \code{"Pi_u"} The superiority measure for unfavorable environments.
\item \code{"R_u"} The rank for Pi_u.
}

\strong{Object is of class \code{performs_ammi}.}
\itemize{
\item \code{"Y"} for raw means (Default).
\item \code{"ipca_ss"} Sum of square for each IPCA.
\item \code{"ipca_ms"} Mean square for each IPCA.
\item \code{"ipca_fval"} F value for each IPCA.
\item \code{"ipca_pval"} P-value for for each IPCA.
\item \code{"ipca_expl"}  Explained sum of square for each IPCA.
\item \code{"ipca_accum"} Accumulated explained sum of square.
}

\strong{Object is of class \code{waas} and \code{waasb}.}
\itemize{
\item \code{"Y"} for raw means (Default).
\item \code{"PCn"} for the scores of the nth Interaction Principal Component Axis (IPCA).
where * \code{n} is the desired IPCA (e.g. \code{"PC3"}).
\item \code{"WAASB"}  The weighted average of the absolute scores.
\item \code{"PctResp"} The rescaled values of the response variable.
\item \code{"PctWAASB"} The rescaled values of the WAASB.
\item \code{"wResp"} The weight for the response variable.
\item \code{"wWAASB"} The weight for the stability.
\item \code{"OrResp"} The ranking regarding the response variable.
\item \code{"OrWAASB"} The ranking regarding the WAASB.
\item \code{"OrPC1"} The ranking regarding the first principal component axix.
\item \code{"WAASBY"} The superiority index WAASBY.
\item \code{"OrWAASBY"} The ranking regarding the superiority index.
}

\strong{Object is of class \code{waasb}
or \code{gamem}.}
\itemize{
\item \code{"HMGV"} For harmonic mean of genotypic values.
\item \code{"RPGV or RPGV_Y"} For relative performance of genotypic values
\item \code{"HMRPGV"} For harmonic mean of relative performance of genotypic values
}

\strong{Object is of class \code{Res_ind}}
\itemize{
\item \code{"blupg"} For genotype's predicted mean.
\item \code{"blupge"} for genotype-vs-environment's predicted mean (only for
objects of class \code{waasb}).
\item \code{"genpar"} Genetic parameters.
\item \code{"lrt"} The statistic for the likelihood-ratio test for random effects.
\item \code{"pval_lrt"}  The p-values for the likelihood-ratio test.
\item \code{"vcomp"} The variance components for random effects.
}
}
\examples{

library(metan)
#################### AMMI model #####################
# Fit an AMMI model for 7 variables.
AMMI <- data_ge2 \%>\%
 performs_ammi(ENV, GEN, REP,
               resp = c(PH, ED, TKW, NKR, CD, CL, CW))

# Sum of squares
get_model_data(AMMI, "ipca_ss")

# Mean squares
get_model_data(AMMI, "ipca_ms")

# Examine the significance (p-value) of the IPCAs
get_model_data(AMMI, "ipca_pval")

# Explained sum of square for each IPCA
get_model_data(AMMI, "ipca_expl")

# Accumulated sum of square
get_model_data(AMMI, "ipca_accum")

### AMMI-based stability statistics ###
# Get the AMMI stability value
AMMI \%>\%
AMMI_indexes() \%>\%
get_model_data("ASV")


#################### WAASB model #####################
# Fitting the WAAS index
AMMI <- waas(data_ge2, ENV, GEN, REP,
             resp = c(PH, ED, TKW, NKR))

# Getting the weighted average of absolute scores
get_model_data(AMMI, what = "WAASB")

# And the rank for the WAASB inde.
get_model_data(AMMI, what = "OrWAASB")


#################### BLUP model #####################
# Fitting a mixed-effect model

# Getting p-values for likelihood-ratio test
get_model_data(blup, what = "pval_lrt")

# Getting the variance components
get_model_data(blup, what = "vcomp")

# Getting the genetic parameters
get_model_data(blup, what = "genpar")

### BLUP-based stability indexes ###
blup \%>\%
Resende_indexes() \%>\%
get_model_data("HMRPGV")

}
\references{
Olivoto, T., A.D.C. L{'{u}}cio, J.A.G. da silva, V.S. Marchioro, V.Q. de
Souza, and E. Jost. 2019a. Mean performance and stability in
multi-environment trials I: Combining features of AMMI and BLUP techniques.
Agron. J.
\href{https://dl.sciencesocieties.org/publications/aj/abstracts/0/0/agronj2019.03.0220}{doi:10.2134/agronj2019.03.0220}

Olivoto, T., A.D.C. L{'{u}}cio, J.A.G. da silva, B.G. Sari, and M.I. Diel.
2019b. Mean performance and stability in multi-environment trials II:
Selection based on multiple traits. Agron. J.
\href{https://dl.sciencesocieties.org/publications/aj/abstracts/0/0/agronj2019.03.0221}{doi:10.2134/agronj2019.03.0221}
}
\author{
Tiago Olivoto \email{tiagoolivoto@gmail.com}
}
