% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{int.effects}
\alias{int.effects}
\title{Dataset with genotype-by-environment interaction effects}
\description{
Dataset with genotype-by-environment interaction effects
}
\author{
Tiago Olivoto \email{tiagoolivoto@gmail.com}
}
\keyword{data}
