% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpcorr.R
\name{lpcor}
\alias{lpcor}
\title{Linear and Partial Correlation Coefficients}
\usage{
lpcor(.data, n = NULL, method = "pearson", verbose = TRUE)
}
\arguments{
\item{.data}{The data to be analyzed. Must be a symmetric correlation matrix
or, a dataframe containing the predictor variables, or an object of class
\code{split_factors}.}

\item{n}{If a correlation matrix is provided, then \code{n} is the number of
objects used to compute the correlation coefficients.}

\item{method}{a character string indicating which correlation coefficient is to be computed.
One of "pearson" (default), "kendall", or "spearman".}

\item{verbose}{If \code{verbose = TRUE} then some results are shown in the
console.}
}
\value{
If a grouping factor is used then a list is returned with the
following values. \item{linear.mat}{The matrix of linear correlation.}
\item{partial.mat}{The matrix of partial correlations.}
\item{results}{Hypotesis testing for each pairwise comparation.}
}
\description{
Estimates the linear and partial correlation coefficients using as input a
data frame or a correlation matrix.
}
\examples{

library(metan)
library(dplyr)
partial1 = lpcor(iris)

# Alternatively using the pipe operator \%>\%
partial2 = iris \%>\% lpcor()

partial3 = cor(iris[1:4]) \%>\%
           lpcor(n = nrow(iris),
                 verbose = FALSE)

partial4 = iris \%>\%
           split_factors(Species) \%>\%
           lpcor()

}
\author{
Tiago Olivoto \email{tiagoolivoto@gmail.com}
}
