% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_sym.R
\name{make_sym}
\alias{make_sym}
\title{Make a symmetric matrix on a triangular matrix}
\usage{
make_sym(.matrix, make = "upper")
}
\arguments{
\item{.matrix}{The upper or lower triangular matrix.}

\item{make}{The triangular to built. Default is \code{"upper"}. In this
case, a symetric matrix will be built based on the values of a lower
triangular values.}
}
\value{
A symmetric matrix.
}
\description{
This function help users to easily make a symmetric matrix using a lower or
an upper triangular matrix.
}
\examples{

m <- matrix(NA,4,4)
m[lower.tri(m)] <- 1:16
m
make_sym(m)

}
\author{
Tiago Olivoto \email{tiagoolivoto@gmail.com}
}
