% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{means_by}
\alias{means_by}
\title{Means by one or more factors}
\usage{
means_by(.data, ...)
}
\arguments{
\item{.data}{A data frame}

\item{...}{One or more categorical variables for grouping the data.}
}
\description{
Computes the mean values for all numeric variables of a data by a factor
}
\examples{
library(metan)
data_ge2 \%>\%
means_by(ENV)
}
