% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metan-package.r
\docType{package}
\name{metan-package}
\alias{metan-package}
\title{Multi-Environment Trial Analysis}
\description{
This package provides useful functions for computing stability
analysis in multi-environment trials (MET) using the most widely-know
parametric and nonparametric methods. A number of vignettes are provided to help the user apply
correctely the functions provide by \pkg{metan}. A complete guide may be
found at https://tiagoolivoto.github.io/metan/
}
