% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metan-package.r
\docType{package}
\name{metan-package}
\alias{metan-package}
\title{Multi-Environment Trial Analysis}
\description{
Provides functions for performing the most
used analyses in the evaluation of multi-environment trials.
This includes stability analysis using widely-known parametric
and nonparametric methods, functions for biometrical analysis such
as path analysis, canonical correlation, partial correlation,
clustering analysis, as well as tools for summarizing and plotting data.
A complete guide may be found at https://tiagoolivoto.github.io/metan/
}
