% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtsi.R
\name{mtsi}
\alias{mtsi}
\title{Multi-trait stability index}
\usage{
mtsi(.data, index = "waasb", SI = 15, mineval = 1, verbose = TRUE)
}
\arguments{
\item{.data}{An object of class \code{waasb} or \code{waas}.}

\item{index}{If \code{index = 'waasb'} (Default) the multitrait index will
be computed considering the stability of genotypes only. If \code{index =
'waasby'} both stability and mean performance are considered. More details
can be seen in \code{\link{waasb}} and \code{\link{waas}} functions.}

\item{SI}{An integer [0-100]. The selection intensity in percentage of the
total number of genotypes.}

\item{mineval}{The minimum value so that an eigenvector is retained in the
factor analysis.}

\item{verbose}{If \code{verbose = TRUE} (Default) then some results are
shown in the console.}
}
\value{
An object of class \code{ge_factanal} with the following items:
\item{data}{The data used to compute the factor analysis.}
\item{cormat}{The correlation matrix among the environments.}
\item{PCA}{The eigenvalues and explained variance.}
\item{FA}{The factor analysis.}
\item{KMO}{The result for the Kaiser-Meyer-Olkin test.}
\item{MSA}{The measure of sampling adequacy for individual variable.}
\item{communalities}{The communalities.}
\item{communalities.mean}{The communalities' mean.}
\item{initial.loadings}{The initial loadings.}
\item{finish.loadings}{The final loadings after varimax rotation.}
\item{canonical.loadings}{The canonical loadings.}
\item{scores.gen}{The scores for genotypes in all retained factors.}
\item{scores.ide}{The scores for the ideotype in all retained factors.}
\item{MTSI}{The multi-trait stability index.}
\item{contri.fac}{The relative contribution of each factor on the MTSI value.
The lower the contribution of a factor, the close of the ideotype the variables in such
factor are.}
\item{sel.dif}{The selection differential for the WAASBY or WAASB index.}
\item{mean.sd}{The mean for the differential selection.}
\item{sel.dif.var}{The selection differential for the variables.}
\item{Selected}{The selected genotypes.}
}
\description{
Computes the multitrait stability index proposed by Olivoto et al. (2019)
}
\examples{

library(metan)

# Based on stability only, for both GY and HM, higher is better
mtsi_model = waasb(data_ge,
                   env = ENV,
                   gen = GEN,
                   rep = REP,
                   resp = c(GY, HM))
mtsi_index = mtsi(mtsi_model)


# Based on mean performance and stability (using pipe operator \%>\%)
# GY: higher is better
# HM: lower is better
mtsi_index2 = data_ge \%>\%
              waasb(ENV, GEN, REP, c(GY, HM), mresp = c(100, 0)) \%>\%
              mtsi(index = 'waasby')

}
\references{
Olivoto, T., A.D.C. L{\'{u}}cio, J.A.G. da silva, B.G. Sari, and M.I. Diel. 2019. Mean performance and stability in multi-environment trials II: Selection based on multiple traits. Agron. J. (in press).
}
\author{
Tiago Olivoto \email{tiagoolivoto@gmail.com}
}
