% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performs_ammi.R
\name{performs_ammi}
\alias{performs_ammi}
\title{Performs Additive Main effects and Multiplicative Interaction}
\usage{
performs_ammi(ENV, GEN, REP, Y)
}
\arguments{
\item{ENV}{The name of the collum that contains the levels of the
environments}

\item{GEN}{The name of the collum that contains the levels of the genotypes}

\item{REP}{The name of the collum that contains the levels of the
replications/blocks}

\item{Y}{The response variable}
}
\value{
\item{ANOVA}{The analysis of variance for the AMMI model.}

\item{analysis}{The principal component analysis}

\item{means}{means of genotype vs environment}

\item{biplot}{escores for genotypes and environments in all the possible
axes.}
}
\description{
Compute the Additive Main effects and Multiplicative interaction. This is a
helper function for other procedures performed in the WAASB package such as
\code{\link{WAAS.AMMI}}, \code{\link{cv_ammi}}, and \code{\link{cv_ammif}}.
}
\examples{

library(metan)
ammi_model = with(data_ge, performs_ammi(ENV, GEN, REP, GY))


}
\author{
Tiago Olivoto \email{tiagoolivoto@gmail.com}
}
