% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performs_ammi.R
\name{performs_ammi}
\alias{performs_ammi}
\title{Additive Main effects and Multiplicative Interaction}
\usage{
performs_ammi(.data, env, gen, rep, resp)
}
\arguments{
\item{.data}{The dataset containing the columns related to Environments,
Genotypes, replication/block and response variable(s).}

\item{env}{The name of the column that contains the levels of the
environments}

\item{gen}{The name of the column that contains the levels of the genotypes}

\item{rep}{The name of the column that contains the levels of the
replications/blocks}

\item{resp}{The response variable}
}
\value{
\item{ANOVA}{The analysis of variance for the AMMI model.}

\item{analysis}{The principal component analysis}

\item{means}{means of genotype vs environment}

\item{biplot}{scores for genotypes and environments in all the possible
axes.}
}
\description{
Compute the Additive Main effects and Multiplicative interaction. This is a
helper function for other procedures performed in the \pkg{metan} package such as
\code{\link{waas}}, \code{\link{cv_ammi}}, and \code{\link{cv_ammif}}.
}
\examples{

library(metan)
ammi_model = performs_ammi(data_ge, ENV, GEN, REP, GY)


}
\author{
Tiago Olivoto \email{tiagoolivoto@gmail.com}
}
