% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.clustering.R
\name{plot.clustering}
\alias{plot.clustering}
\title{Plot an object of class clustering}
\usage{
\method{plot}{clustering}(x, type = "dendrogram", ...)
}
\arguments{
\item{x}{An object of class \code{clustering}}

\item{type}{The type of plot. Must be one of the 'dendrogram' or
'cophenetic'.}

\item{...}{Other arguments passed from the function \code{plot.dendrogram}.}
}
\description{
Plot the multitrait stability index using ggplot-generated graphics. .
}
\author{
Tiago Olivoto \email{tiagoolivoto@gmail.com}
}
