\name{plot.clustering}
\alias{plot.clustering}
\title{Plot an object of class clustering}
\usage{
\method{plot}{clustering}(x, type = "dendrogram", ...)
}

\description{
Plot the multitrait stability index using ggplot-generated graphics. .
}
\arguments{

  \item{x}{An object of class \code{clustering} }

  \item{type}{The type of plot. Must be one of the 'dendrogram' or 'cophenetic'.}


  \item{...}{Other arguments passed from the function \code{plot.dendrogram}.}

}


\author{Tiago Olivoto \email{tiagoolivoto@gmail.com}}


