% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.mtsi.R
\name{plot.mtsi}
\alias{plot.mtsi}
\title{Plot the multi-trait stability index}
\usage{
\method{plot}{mtsi}(x, SI = 15, radar = TRUE, size.point = 2.5,
  col.sel = "red", col.nonsel = "black", ...)
}
\arguments{
\item{x}{An object of class \code{mtsi}}

\item{SI}{An integer [0-100]. The selection intensity in percentage of the
total number of genotypes.}

\item{radar}{Logical argument. If true (default) a radar plot is generated
after using \code{coord_polar()}.}

\item{size.point}{The size of the point in graphic. Defaults to 2.5.}

\item{col.sel}{The colour for selected genotypes.}

\item{col.nonsel}{The colour for nonselected genotypes.}

\item{...}{Other arguments of the function.}
}
\description{
Makes a radar plot showing the multitrait stability index proposed by Olivoto et al. (2019)
}
\examples{
\dontrun{
library(metan)
mtsi_model = waasb(data_ge,
                   env = ENV,
                   gen = GEN,
                   rep = REP,
                   resp = c(GY, HM))

mtsi_index = mtsi(mtsi_model)
plot(mtsi_index)

# Alternatively using the pipe operator \%>\%
library(dplyr)
mtsi_index2 = data_ge \%>\%
              waasb(ENV, GEN, REP, c(GY, HM)) \%>\%
              mtsi()
plot(mtsi_index2)
}


}
\references{
Olivoto, T., A.D.C. L{\'{u}}cio, J.A.G. da silva, B.G. Sari, and M.I. Diel. 2019. Mean performance and stability in multi-environment trials II: Selection based on multiple traits. Agron. J.doi:10.2134/agronj2019.03.0221.
}
\author{
Tiago Olivoto \email{tiagoolivoto@gmail.com}
}
