% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_factlines.R
\name{plot_factlines}
\alias{plot_factlines}
\title{Fast way to create a line plot}
\usage{
plot_factlines(.data, x, y, group, fit, level = 0.95, xlab = NULL,
  ylab = NULL, legend.position = "bottom", grid = FALSE,
  col = TRUE, alpha = 0.2, size.shape = 1.5, size.line = 1,
  size.text = 12, fontfam = "sans", theme = theme_waasb())
}
\arguments{
\item{.data}{The data set}

\item{x}{The variable in data to be shown in the x axis}

\item{y}{The variable in data to be shown in the y axis}

\item{group}{The grouping variable}

\item{fit}{The polinomial degree to use. It must be an integer
between 1 (linear fit) to 4 (fourth-order polynomial regression.),
or a numeric vector with the same length of the variable in \code{group}}

\item{level}{The fonfidence level}

\item{xlab}{The x label}

\item{ylab}{The y label}

\item{col}{The colour to be used in the line plot and points}

\item{alpha}{The alpha for the color in confidence band}

\item{size.shape}{The size for the shape in plot}

\item{size.line}{The size for the line in the plot}

\item{size.text}{The size of the text}

\item{fontfam}{The family of the font text}
}
\description{
Create a graphic to show a fitted line based on numerical
variables and one grouping variable.
}
\seealso{
\code{\link{plot_lines}}, \code{\link{plot_factbars}}
}
