% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resca.R
\name{resca}
\alias{resca}
\title{Rescale a continuous vector to have specified minimum and maximum values}
\usage{
resca(values, new_min = 0, new_max = 100)
}
\arguments{
\item{values}{continuous vector of values to manipulate.}

\item{new_min}{The minimum value of the new scale. Default is 0.}

\item{new_max}{The maximum value of the new scale. Default is 100}
}
\description{
Helper function used in the WAASB package. It rescales a continuous vector
to have specified minimum and maximum values. Missing values are not
allowed.
}
\examples{

resca(1:10)
resca(c(20, 30, 40), new_min = 5, new_max = 10)

}
\author{
Tiago Olivoto \email{tiagoolivoto@gmail.com}
}
