% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.mtsi.R
\name{summary.mtsi}
\alias{summary.mtsi}
\title{Summarise a mtsi object}
\usage{
\method{summary}{mtsi}(object, export = FALSE, file.name = NULL,
  digits = 4, ...)
}
\arguments{
\item{object}{The \code{mtsi} object}

\item{export}{A logical argument. If \code{TRUE|T}, a *.txt file is exported
to the working directory}

\item{file.name}{The name of the file if \code{export = TRUE}}

\item{digits}{The significant digits to be shown.}

\item{...}{Other arguments of the function}
}
\description{
Summary a \code{mtsi} object in two ways. By default, the results are shown
in the R console. The results can also be exported to the directory.
}
\examples{
library(metan)
# Based on stability only
MTSI_MODEL = waasb(data_ge,
                   resp = c(GY, HM),
                   gen = GEN,
                   env = ENV,
                   rep = REP)

MTSI_index = mtsi(MTSI_MODEL)
summary(MTSI_index)


}
\author{
Tiago Olivoto \email{tiagoolivoto@gmail.com}
}
