% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_metan}
\alias{theme_metan}
\alias{theme_metan_minimal}
\title{Personalized theme for ggplot2-based graphics}
\usage{
theme_metan()

theme_metan_minimal()
}
\description{
Two themes that provide plots with a gray background and major
  grids (\code{theme_metan}) or minimalistic theme with half-open frame, white
  background, and no grid (\code{theme_metan_minimal}). For more details see
  \code{\link[ggplot2]{theme}}.
}
\author{
Tiago Olivoto \email{tiagoolivoto@gmail.com}
}
