% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_metan}
\alias{theme_metan}
\alias{theme_metan_minimal}
\title{Personalized theme for ggplot2-based graphics}
\usage{
theme_metan(grid = "none", col.grid = "white", color.background = "gray95")

theme_metan_minimal()
}
\arguments{
\item{grid}{Control the grid lines in plot. Defaults to \code{"both"} (x and
y major grids). Allows also \code{grid = "x"} for grids in x axis only,
\code{grid = "y"} for grid in y axis only, or \code{grid = "none"} for no
grids.}

\item{col.grid}{The color for the grid lines}

\item{color.background}{The color for the panel background.}
}
\description{
Two themes that provide plots with a gray background and major
  grids (\code{theme_metan}) or minimalistic theme with half-open frame, white
  background, and no grid (\code{theme_metan_minimal}). For more details see
  \code{\link[ggplot2]{theme}}.
}
\author{
Tiago Olivoto \email{tiagoolivoto@gmail.com}
}
