\name{WAASB}
\alias{WAASB}
\title{Weighted Average of Absolute Scores for the BLUP'S GxE effects matrix generated by a mixed-effect model}
\usage{
WAASB(data, resp, gen, env, rep, mresp = NULL,
      wresp = NULL, random = "gen", prob = 0.95,
      verbose = TRUE)
}
\description{
Compute the Weighted Average of Absolute Scores for PCA analysis based on the GE-interaction matrix.

}

\arguments{

 \item{data}{The dataset containing the columns related to Environments, Genotypes, replication/block and response variable(s).}

  \item{resp}{The response variable(s). To analyze multiple variables in a single procedure a vector of variables may be used. For example \code{resp = c(var1, var2, var3}.}

  \item{gen}{The name of the column that contains the levels of the genotypes.}

  \item{env}{The name of the column that contains the levels of the environments.}

  \item{rep}{The name of the column that contains the levels of the replications/blocks.}

  \item{mresp}{A numeric vector of the same length of \code{resp}. The \code{mresp} will be the new maximum value after reescaling. By default, all variables in \code{resp} are rescaled so that de maximum value is 100 and the minimum value is 0.}

  \item{wresp}{The weight for the response variable(s) for computing the WAASBY index. Must be a numeric vector of the same length of \code{resp}. Defatul is 50, i.e., equal weights for stability and mean performance.}

  \item{random}{The effects of the model assumed to be random. Default is \code{random = "gen"} (genotype and genotype-vs-environment as random effects. Other values allowed are \code{random = "env"} (environment, genotype-vs-environment and block-within-environment random effects) or \code{random = "all"} all effects except the intercept are assumed to be random effects.}

  \item{prob}{The probability for estimating confidence interval for BLUP's prediction. }

}

\details{
This function compute the weighted average of absolute scores considering all principal component axis from the Singular Value Decomposition (SVD) of the BLUP'S GxE effects matrix generated by a linear mixed-effect model. The main advantage of this procedure in relation to the \link{\code{WAAS.AMMI}} function is that random effects can be included in the model. In addition, unbalanced datasets can also be modeled.
}



\value{
The function returns the results in a list for each analyzed variable. For each variable, the following objects are returned.

\item{individual}{A within-environments ANOVA considering a fixed-effect model.}

\item{model}{A data frame with the response variable, the scores of all Principal Components, the estimates of Weighted Average of Absolute Scores, and WAASY (the index that consider the weights for stability and productivity in the genotype ranking.}

\item{BLUPgen}{The estimated BLUPS for genotypes (If \code{random = "gen"} or \code{random = "all"})}

\item{BLUPenv}{The estimated BLUPS for environments, (If \code{random = "env"} or \code{random = "all"}).}

\item{BLUPge}{The estimated BLUPS of all genotypes in all environments "BLUPij".}

\item{PCA}{The results of Principal Component Analysis with eigenvalues and explained variance of BLUP-interaction matrix.}

\item{MeansGxE}{The phenotypic means of genotypes in the environments, with observed, predicted and OLS residual prediction.}


\item{Details}{A list summarizing the results. The following information are showed. \code{WgtResponse}, the weight for the response variable in estimating WAASB, \code{WgtWAAS} the weight for stability, \code{Ngen} the number of genotypes, \code{Nenv} the number of environments, \code{OVmean} the overall mean, \code{Min} the minimum observed (returning the genotype and environment), \code{Max} the maximum observed, \code{Max} the maximum observed, \code{MinENV} the environment with the lower mean, \code{MaxENV} the environment with the larger mean observed, \code{MinGEN} the genotype with the lower mean,  \code{MaxGEN} the genotype with the larger.}

\item{REML}{A list with the following data frames: \code{fixed}, that contains the ANOVA table with F-tests and p-values for fixed effects (If \code{random = "gen"}); and \code{random} that contains the estimated variance components of the random effects.}

\item{ESTIMATES}{A list with the following values: \code{GEV} the genotype-by-environment variance (and percentage of phenotypic variance); \code{GV} the genotypic variance (and percentage of phenotypic variance); \code{EV} the environmental variance;\code{RV} the residual variance (and percentage of phenotypic variance); \code{FV} the phenotypic variance;  \code{h2g} the heritability of the trait; \code{GEr2} the coefficient of determination of the interaction effects; \code{h2mg} the heritability of the mean; \code{AccuGen} the selective accuracy; \code{rge} the genotype-environment correlation; \code{CVg} the genotypic coefficient of variation; \code{CVr} the residual coefficient of variation; \code{CVratio} the ratio between genotypic and residual coefficient of variation.}

\item{LRT}{The Likelihood Ratio Test for the random effects of the model.}

}

\author{Tiago Olivoto \email{tiagoolivoto@gmail.com}}

\examples{
# Ggenotypes as random effects and equal weights for both response variable and stability

library(METAAB)
model = WAASB(WAASBdata, GY, GEN, ENV, REP)

# Higher weight for stability
model2 = WAASB(WAASBdata, rep = REP, env = ENV, resp = GY, gen = GEN,
               weight.response = 70, weight.WAAS = 30)

# A random-effect model
model3 = WAASB(WAASBdata, GY, GEN, ENV, REP, random = "all")

}


