% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waasb.R
\name{waasb}
\alias{waasb}
\title{Weighted Average of Absolute Scores}
\usage{
waasb(.data, env, gen, rep, resp, mresp = NULL, wresp = NULL,
  random = "gen", prob = 0.05, ind_anova = TRUE, verbose = TRUE)
}
\arguments{
\item{.data}{The dataset containing the columns related to Environments,
Genotypes, replication/block and response variable(s).}

\item{env}{The name of the column that contains the levels of the
environments.}

\item{gen}{The name of the column that contains the levels of the genotypes.}

\item{rep}{The name of the column that contains the levels of the
replications/blocks.}

\item{resp}{The response variable(s). To analyze multiple variables in a
single procedure a vector of variables may be used. For example \code{resp =
c(var1, var2, var3)}.}

\item{mresp}{A numeric vector of the same length of \code{resp}. The
\code{mresp} will be the new maximum value after reescaling. By default, all
variables in \code{resp} are rescaled so that de maximum value is 100 and
the minimum value is 0.}

\item{wresp}{The weight for the response variable(s) for computing the
WAASBY index. Must be a numeric vector of the same length of \code{resp}.
Defatul is 50, i.e., equal weights for stability and mean performance.}

\item{random}{The effects of the model assumed to be random. Default is
\code{random = "gen"} (genotype and genotype-vs-environment as random
effects. Other values allowed are \code{random = "env"} (environment,
genotype-vs-environment and block-within-environment random effects) or
\code{random = "all"} all effects except the intercept are assumed to be
random effects.}

\item{prob}{The probability for estimating confidence interval for BLUP's
prediction.}

\item{ind_anova}{Logical argument set to \code{TRUE}. If \code{FALSE} the
within-environment ANOVA is not performed.}

\item{verbose}{Logical argument. If \code{verbose = FALSE} the code are run
silently.}
}
\value{
An object of class \code{waasb} with the following items for each variable:
\item{individual}{A within-environments ANOVA considering a fixed-effect
model.}

\item{fixed}{Test for fixed effects.}

\item{random}{Variance components for random effects.}

\item{LRT}{The Likelihood Ratio Test for the random effects.}

\item{model}{A data frame with the response variable, the scores of all
Principal Components, the estimates of Weighted Average of Absolute Scores,
and WAASY (the index that consider the weights for stability and
productivity in the genotype ranking.}

\item{blupGEN}{The estimated BLUPS for genotypes (If \code{random = "gen"}
or \code{random = "all"})}

\item{BLUPenv}{The estimated BLUPS for environments, (If \code{random =
"env"} or \code{random = "all"}).}

\item{BLUPge}{The estimated BLUPS of all genotypes in all environments
"BLUPij".}

\item{PCA}{The results of Principal Component Analysis with eigenvalues and
explained variance of BLUP-interaction matrix.}

\item{MeansGxE}{The phenotypic means of genotypes in the environments, with
observed, predicted and OLS residual prediction.}

\item{Details}{A list summarizing the results. The following information are
showed. \code{WgtResponse}, the weight for the response variable in
estimating WAASB, \code{WgtWAAS} the weight for stability, \code{Ngen} the
number of genotypes, \code{Nenv} the number of environments, \code{OVmean}
the overall mean, \code{Min} the minimum observed (returning the genotype
and environment), \code{Max} the maximum observed, \code{Max} the maximum
observed, \code{MinENV} the environment with the lower mean, \code{MaxENV}
the environment with the larger mean observed, \code{MinGEN} the genotype
with the lower mean, \code{MaxGEN} the genotype with the larger.}

\item{ESTIMATES}{A list with the following values: \code{GEV} the
genotype-by-environment variance (and percentage of phenotypic variance);
\code{GV} the genotypic variance (and percentage of phenotypic variance);
\code{EV} the environmental variance;\code{RV} the residual variance (and
percentage of phenotypic variance); \code{FV} the phenotypic variance;
\code{h2g} the heritability of the trait; \code{GEr2} the coefficient of
determination of the interaction effects; \code{h2mg} the heritability of
the mean; \code{AccuGen} the selective accuracy; \code{rge} the
genotype-environment correlation; \code{CVg} the genotypic coefficient of
variation; \code{CVr} the residual coefficient of variation; \code{CVratio}
the ratio between genotypic and residual coefficient of variation.}
\item{residuals}{The residuals of the model.}
}
\description{
Compute the Weighted Average of Absolute Scores for quantifying the
stability in multienvironment trials using mixed-effect models (Olivoto et al., 2019).
}
\details{
This function compute the weighted average of absolute scores considering
all principal component axis from the Singular Value Decomposition (SVD) of
the BLUP'S GxE effects matrix generated by a linear mixed-effect model.
}
\examples{
\dontrun{
library(metan)

# Genotypes as random effects and equal weights for both

# response variable and stability

model <-  waasb(data_ge,
               env = ENV,
               gen = GEN,
               rep = REP,
               resp = GY,
               wresp = 70)

# Higher weight for response variable

model2 <- waasb(data_ge,
               env = ENV,
               gen = GEN,
               rep = REP,
               resp = GY,
               wresp = 65)

# Environment as random effects analyzing more than one variables
# considering that smaller values of HM are better and higher
# values of GY are better, assigning a larger weight for the GY
# and a smaller weight for HM when computing WAASBY index.

model3 <- waasb(data_ge,
                env = ENV,
                gen = GEN,
                rep = REP,
                resp = c(GY, HM),
                random = "env",
                mresp = c(100, 0),
                wresp = c(60, 40))
}

}
\references{
Olivoto, T., A.D.C. L{\'{u}}cio, J.A.G. da silva, V.S. Marchioro,
 V.Q. de Souza, and E. Jost. 2019. Mean performance and stability in multi-environment
  trials I: Combining features of AMMI and BLUP techniques. Agron. J. (in press).
}
\seealso{
\code{\link{waas}}
}
\author{
Tiago Olivoto \email{tiagoolivoto@gmail.com}
}
